### 2023-12-04 #6122 @editorialbot
[PRE REVIEW]: RidePy: A fast and modular framework for simulating ridepooling systems
**Submitting author:** @fxjung (<a href="http://orcid.org/0000-0002-8689-3713">Felix Jung</a>)
**Repository:** https://github.com/PhysicsOfMobility/ridepy
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1.post4
