### 2023-12-04 #6122 @danielskatz
@openjournals/dev - Can you help me understand this problem? I don't see anything in the log that tells me what failed and why, at the level of the paper.
