### 2023-12-05 #6122 @xuanxu
It looks like the [date](https://github.com/PhysicsOfMobility/ridepy/blob/joss/joss/paper.md?plain=1#L19) in the paper's metadata is wrongly formatted. [This PR](https://github.com/PhysicsOfMobility/ridepy/pull/227) should fix the error.
