+++ joss/paper.bib
@@ -122,6 +122,7 @@
   year = {2019},
   journal = {Transp Res Policy Pr.},
   volume = {119},
+  doi = {10.1016/j.tra.2018.10.028},
   pages = {15--28},
   file = {/home/jung/Zotero/storage/MVQ7HVM9/Herminghaus - 2019 - Mean field theory of demand responsive ride poolin.pdf;/home/jung/Zotero/storage/239T62BM/S0965856417316038.html}
 }
@@ -144,6 +145,7 @@
   author = {Bray, T.},
   year = {2017},
   month = dec,
+  doi = {10.17487/RFC4180},
   url = {https://www.rfc-editor.org/rfc/rfc4180.txt},
   langid = {english}
 }
