### 2023-12-05 #6122 @fxjung
> 👋 @fxjung - While we get started with finding an editor, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Thanks for pointing out the missing DOIs! I've added two of them. The third reference missing a DOI is `pytest`, which I cited according to their [citation doc file](https://github.com/pytest-dev/pytest/blob/main/CITATION). Unfortunately, there doesn't seem to be a DOI associated with it. The DOI suggested by the bot points to a book chapter which indeed is related to `pytest`, but is not associated with the official `pytest` repository.
