### 2024-01-18 #6241 @editorialbot
[REVIEW]: RidePy: A fast and modular framework for simulating ridepooling systems
**Submitting author:** @fxjung (<a href="http://orcid.org/0000-0002-8689-3713">Felix Jung</a>)
**Repository:** https://github.com/PhysicsOfMobility/ridepy
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.7
