### 2024-01-23 #6241 @xoolive
Dear @fxjung,

Before I start trying to go into the code, I am having serious concerns about the quality of the paper and of the documentation. I entered your project from these two elements, and I am just confused. On both sides, I am having a hard time understanding what the library is about: I get the context, but not really what the library would provide me as an end user. It talks about simulation but then... I am a bit lost.

As an example, you start your [documentation](https://ridepy.org/overview.html#the-setting) by saying what ridepy does **not** do, this is not the most helpful way to catch your potential users.

To be honest, I do not plan to put more efforts in the review until those two key elements are fully clarified.

Also, as a minor comment that would come back later: I do not recommend inflating your bibliography with citations keys for programming languages and widely adopted libraries (such as Pandas). It is commonly accepted that software papers do not have as many references as standard academic/methodological papers.
