+++ doc/overview.rst
@@ -3,7 +3,12 @@ Overview
 
 .. highlight:: python
 
-Here we describe the design of ridepy, and how to get started with using
+
+RidePy is a Python library that runs mobility simulations, in particular it is able to simulate on-demand transit services such as ridehailing and ridepooling. It provides an interface that allows researchers to replicate the operation of such transport systems and thus to study their properties and behavior without actually transporting someone in the real world.
+
+Getting started with RidePy is easy: The user just has to choose a way of generating the virtual customers' requests for transportation and set up a transportation service by specifying its characteristics (e.g., number of vehicles, dispatching algorithm, additional constraints). RidePy will simulate the operation of the service and log all events, e.g., pick-up of a customer or submission of a new request. After the simulation has finished, RidePy can additionally be used to retrieve structured information about the details of the simulation run.
+
+In this chapter we describe the design of RidePy, and how to get started with using
 it to run simulations. We also describe how to specify new dispatching
 algorithms and new transport spaces.
 
@@ -11,14 +16,14 @@ algorithms and new transport spaces.
 
 The setting
 -----------
-ridepy does *not* do agent-based simulations. Rather, it starts with a set of
+RidePy does *not* do agent-based simulations. Rather, it starts with a set of
 *transportation requests* (denoting the desire of an individual to be transported from A
 to B within specified time windows), and a *dispatcher* (an algorithm that determines
 which vehicle should service which requests and in which order). Then it simply
 simulates these requests arriving in the system, being picked up and delivered. Requests
 that cannot be delivered within the specified time windows are *rejected*.
 
-ridepy makes it easy to experiment with different dispatching algorithms,
+RidePy makes it easy to experiment with different dispatching algorithms,
 spatiotemporal request densities, fleet sizes and transport spaces (2-D plane, different
 graphs). It comes with an `analytics` module that computes from the simulation output
 various metrics like the statistical distributions of occupancies, waiting times and
@@ -30,7 +35,7 @@ quantitative study, we include two powerful ways of speeding up the simulation:
 Using parallelism:
    The dispatcher interface prescribes computing a *cost* for serving a request with a
    certain vehicle. Then the vehicle with the minimum cost is chosen. Since this is an
-   "embarassingly parallelizable" operation, ridepy provides two parallel ways of
+   "embarassingly parallelizable" operation, RidePy provides two parallel ways of
    computing it, out of the box:
 
    - ``multiprocessing``,
@@ -131,7 +136,7 @@ delivered.
 
 Using parallelism
 -----------------
-Running ridepy in a multi-node OpenMPI cluster is as simple as replacing
+Running RidePy in a multi-node OpenMPI cluster is as simple as replacing
 :class:`SlowSimpleFleetState <fleet_state.SlowSimpleFleetState>` with
 :class:`MPIFuturesFleetState <fleet_state.MPIFuturesFleetState>`:
 
