### 2024-01-25 fabfb32
removed references for common languages and python libraries as per reviewer request (@xoolive)
