+++ doc/overview.rst
@@ -3,33 +3,21 @@ Overview
 
 .. highlight:: python
 
-RidePy is a Python library that runs mobility simulations; in particular, it is able to simulate on-demand transit services such as ridehailing and ridepooling. It provides an interface that allows researchers to replicate the operation of such transport systems and thus study their properties and behavior without actually transporting someone in the real world.
+RidePy is a Python library that performs mobility simulations; in particular, it is able to simulate on-demand transit services such as ride hailing (taxicab-like operations) and ridepooling (trips of similar origin, destination, and time are served by the same vehicle). RidePy provides an interface that allows researchers to replicate the operation of such transport systems and thus study their properties and behavior without actually transporting passengers in the real world.
 
-Getting started with RidePy is easy: The user just has to choose a way of generating the virtual customers' requests for transportation and set up a transportation service by specifying its characteristics (e.g., number of vehicles, dispatching algorithm, additional constraints). RidePy will simulate the operation of the service and log all events, e.g., the pick-up of a customer or the submission of a new request. After the simulation has finished, RidePy can additionally be used to retrieve structured information about the details of the simulation run.
+Getting started with RidePy is easy: The user just has to choose a way of generating the virtual customers' requests for transportation and set up a transportation service by specifying its characteristics (e.g., number of vehicles, dispatching algorithm, additional constraints). RidePy will simulate the operation of the service and log all events, e.g., the pick-up of a customer or the submission of a new request. After the simulation has finished, RidePy can additionally provide structured information about the details of the simulation run. These data can then be analyzed further by the user to gain insights into the behavior of the simulated service.
 
-In this chapter, we describe the design of RidePy, and how to get started with using
-it to run simulations. We also describe how to specify new dispatching
-algorithms and new transport spaces.
-
+In this chapter, we describe the design of RidePy, and how to get started using it to run simulations. We also describe how to specify new dispatching algorithms and new transport spaces.
 
 
 The setting
 -----------
-RidePy does *not* do agent-based simulations. Rather, it starts with a set of
-*transportation requests* (denoting the desire of an individual to be transported from A
-to B within specified time windows), and a *dispatcher* (an algorithm that determines
-which vehicle should service which requests and in which order). Then it simply
-simulates these requests arriving in the system, being picked up and delivered. Requests
-that cannot be delivered within the specified time windows are *rejected*.
 
-RidePy makes it easy to experiment with different dispatching algorithms,
-spatiotemporal request densities, fleet sizes and transport spaces (2-D plane, different
-graphs). It comes with an `analytics` module that computes from the simulation output
-various metrics like the statistical distributions of occupancies, waiting times and
-detours.
+Rather than explicitly modeling the behavior of individuals (classic agent-based simulations), RidePy instead starts with a set of *transportation requests* (denoting the desire of an individual to be transported from A to B within a specified time window), and a *dispatcher* (an algorithm that determines which vehicle should service which requests and in which order). Then it simply simulates these requests arriving in the system, being picked up, and being delivered. Requests that cannot be delivered within the specified time windows are *rejected*.
 
-Since the ability to simulate many *requests* and many *vehicles* is important for any
-quantitative study, we include two powerful ways of speeding up the simulation:
[…]
