### 2024-01-25 551a60e
moved things from readme to glossary
