### 2024-01-25 e49d3b0
extended documentation (#229)
* extended overview section

* typos

* typos

* further enhancement of overview

* added glossary

* moved things from readme to glossary

* mention notebooks
