+++ .github/workflows/deploy-testpypi.yml
@@ -25,11 +25,11 @@ jobs:
     environment: pypi
     needs: [sdist, wheel, test, doc]
     steps:
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
           name: sdist
           path: dist
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
           name: wheel-manylinux
           path: dist
