+++ .github/workflows/build-wheel.yml
@@ -8,11 +8,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout repository and submodules
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: Checkout submodules
         run: git submodule update --init --recursive
       - name: Set up Python 3.9
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: '3.9'
       - name: Cache python environment
@@ -33,7 +33,7 @@ jobs:
 #          pre-build-command: 'sh pre-build-script.sh'
 #          package-path: 'my_project'
 #          pip-wheel-args: '-w ./dist --no-deps'
-      - uses: actions/upload-artifact@v3
+      - uses: actions/upload-artifact@v4
         with:
           name: wheel-manylinux
           path: dist/*-manylinux*.whl
