+++ .github/workflows/deploy-pypi.yml
@@ -23,11 +23,11 @@ jobs:
     environment: pypi
     needs: [sdist, wheel, test, deploy-doc]
     steps:
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
           name: sdist
           path: dist
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
           name: wheel-manylinux
           path: dist
