### 2024-01-25 #6241 @fxjung
> Dear @fxjung,
>
> Before I start trying to go into the code, I am having serious concerns about the quality of the paper and of the documentation. I entered your project from these two elements, and I am just confused. On both sides, I am having a hard time understanding what the library is about: I get the context, but not really what the library would provide me as an end user. It talks about simulation but then... I am a bit lost.
>
> As an example, you start your [documentation](https://ridepy.org/overview.html#the-setting) by saying what ridepy does **not** do, this is not the most helpful way to catch your potential users.
>
> To be honest, I do not plan to put more efforts in the review until those two key elements are fully clarified.
>
> Also, as a minor comment that would come back later: I do not recommend inflating your bibliography with citations keys for programming languages and widely adopted libraries (such as Pandas). It is commonly accepted that software papers do not have as many references as standard academic/methodological papers.

Dear @xoolive,

Thank you for reviewing our paper and for your comments.

I am sorry for any confusion regarding the purpose of the library. I have now tried to make the description of the functionality more accessible/high-level in the [overview section](https://ridepy.org/overview). Furthermore, the documentation now additionally mentions the introductory jupyter notebooks under [Quickstart](https://ridepy.org/overview.html#quickstart). Finally, I have updated the README, moving possibly confusing internals at the end to a separate glossary chapter in the documentation and highlighting the broad structure of the documentation at the beginning. I hope this addresses your concerns.

Unfortunately, I think I do not fully understand your criticism regarding the manuscript. We have tried to point out as clear as possible what the library does and what it's useful for both in the *Statement of need* and *Philosophy of usage* sections. It would be great if you could elaborate a bit further on what exactly may be missing here.

Coming from physics, I assumed it would be sensible to cite even popular languages and libraries. As you implied that this is uncommon in software literature, I have removed them accordingly.

Regarding your comment on automated tests: The project is already extensively tested using pytest. The tests are located in the `test` folder as configured in `pytest.ini`. These tests are automatically run using GitHub actions and may also be run locally by executing `pytest` in the repository root. Manually testing whether the software works is also possibly using the example notebooks in the `notebooks` directory.

   Felix
