### 2024-01-26 e5e13ab
fixed analytics bug: median stoplist length
