+++ src/ridepy/util/analytics/__init__.py
@@ -467,11 +467,11 @@ def _add_insertion_stats_to_stoplist_dataframe(*, reqs, stops, space) -> pd.Data
     reqs
         requests  DataFrame
     stops
-        stops DataFrame missing stop locations for non-internal stops
+        stops DataFrame
 
     Returns
     -------
-    stoplist DataFrame with added stop locations indexed by `vehicle_id` and `timestamp`
+    stoplist DataFrame
     """
 
     stops = pd.merge(
@@ -606,7 +606,7 @@ def _add_insertion_stats_to_stoplist_dataframe(*, reqs, stops, space) -> pd.Data
         return res
 
     stops = stops.merge(
-        actual_stops.groupby("vehicle_id").apply(
+        actual_stops.groupby("vehicle_id", group_keys=False).apply(
             lambda df: df.apply(
                 ft.partial(_properties_at_time, full_sl=df, scope="vehicle"),
                 axis=1,
@@ -662,12 +662,16 @@ def get_stops_and_requests_from_events_dataframe(
         evs=events_df, stops=stops_df, space=space
     )
 
-    try:
-        stops_df = _add_locations_to_stoplist_dataframe(
-            reqs=requests_df, stops=stops_df, space=space
-        )
-    except KeyError:  # TODO document this
-        pass
+    # try:
+    stops_df = _add_locations_to_stoplist_dataframe(
+        reqs=requests_df, stops=stops_df, space=space
+    )
+    # stops_df = _add_insertion_stats_to_stoplist_dataframe(
+    #     reqs=requests_df, stops=stops_df, space=space
+    # )
+    # except KeyError:  # TODO document this
+
+    # pass
 
     return stops_df, requests_df
 
@@ -778,7 +782,7 @@ def get_vehicle_quantities(stops: pd.DataFrame, requests: pd.DataFrame) -> pd.Da
     """
     Compute various quantities aggregated **per vehicle**.
 
-    Currently the following observables are returned:
+    Currently, the following observables are returned:
 
     - avg_occupancy
     - avg_segment_dist
@@ -843,31 +847,7 @@ def get_vehicle_quantities(stops: pd.DataFrame, requests: pd.DataFrame) -> pd.Da
     efficiency_dist = total_direct_dist / total_dist_driven
     efficiency_time = total_direct_time / total_time_driven
 
-    avg_system_stoplist_length_service_time = stops.groupby("vehicle_id").apply(
-        lambda gdf: (
-            gdf["system_stoplist_length_service_time"] * gdf["state_duration"]
-        ).sum()
-        / gdf["state_duration"].sum()
-    )
-    avg_system_stoplist_length_submission_time = stops.groupby("vehicle_id").apply(
-        lambda gdf: (
-            gdf["system_stoplist_length_submission_time"] * gdf["state_duration"]
-        ).sum()
-        / gdf["state_duration"].sum()
-    )
-
-    avg_stoplist_length_service_time = stops.groupby("vehicle_id").apply(
-        lambda gdf: (gdf["stoplist_length_service_time"] * gdf["state_duration"]).sum()
-        / gdf["state_duration"].sum()
-    )
-    avg_stoplist_length_submission_time = stops.groupby("vehicle_id").apply(
-        lambda gdf: (
-            gdf["stoplist_length_submission_time"] * gdf["state_duration"]
-        ).sum()
-        / gdf["state_duration"].sum()
-    )
-
-    return pd.DataFrame(
+    res = pd.DataFrame(
         dict(
             avg_occupancy=avg_occupancy,
             avg_segment_dist=avg_segment_dist,
@@ -880,12 +860,45 @@ def get_vehicle_quantities(stops: pd.DataFrame, requests: pd.DataFrame) -> pd.Da
[…]
