### 2024-01-26 3ec4daa
fixed analytics bug: median stoplist length (#231)
