+++ test/test_analytics.py
@@ -901,7 +901,7 @@ def test_get_stops_and_requests_and_get_quantities():
             "efficiency_dist": (0.3 + 0.1 + 1) / (0.1 + 0.1 + 0.1 + 1 + 1),
             "efficiency_time": (0.3 + 0.1 + 1) / (0.1 + 0.1 + 0.1 + 1 + 1),
             "avg_waiting_time": (0 + 0.1 + 1) / 3,
-            "median_stoplist_length": 2.0,
+            "median_stoplist_length": np.median([3, 2, 1, 0, 0, 2, 1, 0, 0, 0]),
             "rejection_ratio": 0.25,
             "avg_detour": 1.0,
             "avg_system_stoplist_length_service_time": 1.4666666666666668,
@@ -909,6 +909,7 @@ def test_get_stops_and_requests_and_get_quantities():
             "avg_stoplist_length_service_time": 0.13333333333333333,
             "avg_stoplist_length_submission_time": 1.3333333333333333,
         }
+
         assert get_system_quantities(stops, requests) == expected_system_quantities
 
         plot_occupancy_hist(stops)
