### 2024-01-26 #6241 @xoolive
Thank you @fxjung

I am sorry I spoke too fast about the tests, I just missed them :fearful: I will have a look!

To be honest, I still recommend improving the website for the documentation. Avoid twisting the user into opening jupyter notebooks (I never do), but you can render them with sphynx, there's a plugin for that. Maybe it can help demonstrating the capabilities of the library.

I am not expecting a perfect documentation, it just takes too much time to write an excellent on, but the fact is that I am missing the point there. [This](https://documentation.divio.com/) is my favourite guide for documentation by the way.
