### 2024-01-26 e0115f1
restructured analytics module
