### 2024-01-26 602f13c
restructured analytics module (#233)
