+++ src/ridepy/util/analytics/__init__.py
@@ -1,1090 +1,6 @@
-from collections import defaultdict
-
-from typing import Iterable, List, Optional, Dict, Any
-
-import functools as ft
-import numpy as np
-import pandas as pd
-
-from ridepy.data_structures import (
-    TransportSpace,
+from ridepy.util.analytics.events import (
+    get_stops_and_requests,
+    get_stops_and_requests_from_events_dataframe,
 )
-from ridepy.events import VehicleStateEndEvent, VehicleStateBeginEvent
-
-
-def _create_events_dataframe(events: Iterable[dict]) -> pd.DataFrame:
-    """
-    Create a DataFrame of all logged events with their properties at columns.
-
-    The schema of the returned DataFrame is the following, where
-    the index is an unnamed integer range index.
-
-    .. code-block::
-
-        Column                   Dtype
-        ------                   -----
-        request_id               int64
-        timestamp                float64
-        vehicle_id               float64
-        event_type               object
-        location                 Union[int, float64, Tuple[float64]]
-        origin                   Union[int, float64, Tuple[float64]]
-        destination              Union[int, float64, Tuple[float64]]
-        pickup_timewindow_min    float64
-        pickup_timewindow_max    float64
-        delivery_timewindow_min  float64
-        delivery_timewindow_max  float64
-
-    Parameters
-    ----------
-    events
-
-    Returns
-    -------
-    events DataFrame, indexed by integer range
-    """
-    return pd.DataFrame(events)
-
-
-def _create_stoplist_dataframe(*, evs: pd.DataFrame) -> pd.DataFrame:
-    """
-    Creates a DataFrame containing the stoplists of the transporters.
-    The location is only returned for the internal stops relating to
-    VehicleStateBeginEvent and VehicleStateEndEvent. For the rest of
-    the stops, location is nan.
-
-    The schema of the returned DataFrame is the following, where
-    `vehicle_id` and `timestamp` are MultiIndex levels.
-
-    .. code-block::
-
-        Column           Dtype
-        ------           -----
-        vehicle_id       float64
-        timestamp        float64
-        delta_occupancy  float64
-        request_id       object
-        state_duration   float64
-        occupancy        float64
-        location         Union[float64, int, Tuple[float64]]
-
-    Parameters
-    ----------
-    evs
-        events DataFrame
-
-    Returns
-    -------
-    stoplist DataFrame indexed by `vehicle_id` and `timestamp`
-    """
-    # Create the initial dataframe: Select only events of type
-    # `PickupEvent` and `DeliveryEvent`, along with their vehicle_id, timestamp,
-    # and request_id
-    stops = evs[
-        evs["event_type"].isin(
-            [
-                "PickupEvent",
-                "DeliveryEvent",
-                "VehicleStateBeginEvent",
-                "VehicleStateEndEvent",
-            ]
-        )
-    ][["vehicle_id", "timestamp", "event_type", "request_id", "location"]]
-
-    # Translate PickupEvent and DeliveryEvent into an occupancy delta.
-    # NOTE: This assumes occupancy delta of +1/-1, i.e. only single-customer requests.
-    #       If the simulator should allow for multi-customer requests in the future,
-    #       this must be changed.
-    #       See also [issue #45](https://github.com/PhysicsOfMobility/ridepy/issues/45)
-    stops["delta_occupancy"] = stops["event_type"].map(
-        defaultdict(float, PickupEvent=1, DeliveryEvent=-1)
-    )
-
-    # ... and drop the event_type as pickup/dropoff is now signified through delta_occupancy
-    stops.drop("event_type", axis=1, inplace=True)
-
-    stops.sort_values(["vehicle_id", "timestamp", "request_id"], inplace=True)
-
-    # Fix the stop order. The begin and end stops have identical timestamps as
[…]
