### 2024-01-29 53881c5
added load computation to analytics module
