+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 -   repo: https://github.com/python/black
-    rev: 23.1.0
+    rev: 24.1.1
     hooks:
     - id: black
       language_version: python3.9
