### 2024-01-30 2d9ae2c
added additional analytics module glossary
