+++ src/ridepy/util/spaces.py
@@ -548,14 +548,10 @@ class ContinuousGraph(Graph):
         """coordinates shall consist of triples (u, v, dist_to_dest)"""
         ...
 
-    def t(self, u, v):
-        ...
+    def t(self, u, v): ...
 
-    def interp_dist(self, u, w, time_to_dest):
-        ...
+    def interp_dist(self, u, w, time_to_dest): ...
 
-    def interp_time(self, u, v, time_to_dest):
-        ...
+    def interp_time(self, u, v, time_to_dest): ...
 
-    def random_point(self):
-        ...
+    def random_point(self): ...
