### 2024-01-30 266ec96
simulation set bugfix and enhancement
