+++ test/test_space.py
@@ -18,11 +18,13 @@ from ridepy.util.spaces import (
     Euclidean2D,
     Graph,
     DiGraph,
+    Manhattan2D,
 )
 from ridepy.util.spaces_cython import (
     Euclidean2D as CyEuclidean2D,
     Manhattan2D as CyManhattan2D,
     Graph as CyGraph,
+    Grid2D as CyGrid2D,
 )
 
 from ridepy.extras.spaces import (
@@ -54,6 +56,13 @@ def test_Euclidean2D():
     assert space.d((0, 0), (0, 1)) == 1.0
     assert space.d((0, 0), (0, 0)) == 0.0
     assert space.d((0, 0), (1, 1)) == m.sqrt(2)
+
+
+def test_Manhattan2D():
+    space = Manhattan2D()
+    assert space.d((0, 0), (0, 1)) == 1.0
+    assert space.d((0, 0), (0, 0)) == 0.0
+    assert space.d((0, 0), (1, 1)) == 2
 
 
 def test_Euclidean2D_smart_vectorize():
@@ -147,11 +156,147 @@ def test_CyEuclidean2D():
     assert space.d((0, 0), (1, 1)) == m.sqrt(2)
 
 
-def test_Manhattan2D():
+def test_CyManhattan2D():
     space = CyManhattan2D()
     assert space.d((0, 0), (0, 1)) == 1.0
     assert space.d((0, 0), (0, 0)) == 0.0
     assert space.d((0, 0), (1, 1)) == 2
+
+
+def test_CyGrid2D():
+    space = CyGrid2D()
+
+    assert space.d((0, 0), (0, 1)) == 1.0
+    assert space.d((0, 0), (0, 0)) == 0.0
+    assert space.d((0, 0), (1, 1)) == 2
+    assert space.d((0, 0), (0, 0.1)) == 0.1
+
+
+def test_CyGrid2D_velocity():
+    space = CyGrid2D()
+
+    assert space.t((0, 0), (0, 1)) == 1.0
+    assert space.t((0, 0), (0, 0)) == 0.0
+    assert space.t((0, 0), (1, 1)) == 2
+    assert space.t((0, 0), (0, 0.1)) == 0.1
+
+    space = CyGrid2D(velocity=2)
+
+    assert space.t((0, 0), (0, 1)) == 0.5
+    assert space.t((0, 0), (0, 0)) == 0.0
+    assert space.t((0, 0), (1, 1)) == 1
+    assert space.t((0, 0), (0, 0.1)) == 0.05
+
+
+def test_CyGrid2D_interpolation():
+
+    def assert_interpolation_equal(a, b):
+        assert a[0][0] == pytest.approx(b[0][0])
+        assert a[0][1] == pytest.approx(b[0][1])
+        assert a[1] == pytest.approx(b[1])
+
+    space = CyGrid2D()
+
+    assert_interpolation_equal(space.interp_dist((0, 0), (0, 1), 1.0), ((0, 0), 0))
+    # assert space.interp_dist((0, 1), (0, 1), 1.0) == ((1, 0), 0) # this is undefined behavior
+
+    # ========================================
+
+    # X---X   X   X   X
+    #     |
+    # X   X   X   X   X
+    #
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((0, 0), (1, 1), 1.0), ((0, 1), 0))
+
+    # X   X   X   X   X
+    # |
+    # X---X   X   X   X
+    #
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((1, 1), (0, 0), 1.0), ((1, 0), 0))
+
+    # ========================================
+
+    # X---X   X   X   X
+    #     |
+    # X   X   X   X   X
+    #     |
[…]
