+++ test/test_space.py
@@ -203,7 +203,7 @@ def test_CyGrid2D_interpolation():
     # ========================================
 
     # X---X   X   X   X
-    #     |
+    #     V
     # X   X   X   X   X
     #
     # X   X   X   X   X
@@ -211,7 +211,7 @@ def test_CyGrid2D_interpolation():
     assert_interpolation_equal(space.interp_dist((0, 0), (1, 1), 1.0), ((0, 1), 0))
 
     # X   X   X   X   X
-    # |
+    # A
     # X---X   X   X   X
     #
     # X   X   X   X   X
@@ -223,7 +223,7 @@ def test_CyGrid2D_interpolation():
     # X---X   X   X   X
     #     |
     # X   X   X   X   X
-    #     |
+    #     V
     # X   X   X   X   X
 
     assert_interpolation_equal(space.interp_dist((0, 0), (2, 1), 0.6), ((2, 1), 0.6))
@@ -232,7 +232,7 @@ def test_CyGrid2D_interpolation():
     assert_interpolation_equal(space.interp_dist((0, 0), (2, 1), 2.6), ((0, 1), 0.6))
 
     # X   X   X   X   X
-    # |
+    # A
     # X   X   X   X   X
     # |
     # X---X   X   X   X
@@ -245,7 +245,7 @@ def test_CyGrid2D_interpolation():
     # ========================================
 
     # X---X   X   X   X
-    # |
+    # V
     # X   X   X   X   X
     #
     # X   X   X   X   X
@@ -253,7 +253,7 @@ def test_CyGrid2D_interpolation():
     assert_interpolation_equal(space.interp_dist((1, 0), (0, 1), 1.0), ((0, 0), 0))
 
     # X   X   X   X   X
-    #     |
+    #     A
     # X---X   X   X   X
     #
     # X   X   X   X   X
@@ -262,7 +262,7 @@ def test_CyGrid2D_interpolation():
 
     # ========================================
 
-    # X---X   X   X   X
+    # X-->X   X   X   X
     # |
     # X   X   X   X   X
     # |
@@ -272,14 +272,14 @@ def test_CyGrid2D_interpolation():
     assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 1.0), ((0, 0), 0))
     assert_interpolation_equal(
         space.interp_dist((2, 0), (0, 1), 1.6), ((0, 0), 0.6)
-    )  # this is broken
+    )  # this is broken (returns ((1, 1), 0.6), which is incompatible with the previous result)
     assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 2.6), ((1, 0), 0.6))
 
     # X   X   X   X   X
     #     |
     # X   X   X   X   X
     #     |
-    # X---X   X   X   X
+    # X<--X   X   X   X
 
     assert_interpolation_equal(space.interp_dist((0, 1), (2, 0), 0.6), ((2, 0), 0.6))
     assert_interpolation_equal(space.interp_dist((0, 1), (2, 0), 1.0), ((2, 1), 0))
