+++ src/ridepy/util/spaces_cython/__init__.py
@@ -1,1 +1,1 @@
-from .spaces import Euclidean2D, Manhattan2D, Grid2D, Graph, TransportSpace
+from .spaces import Euclidean2D, Manhattan2D, Grid2D, Grid2D_QM, Graph, TransportSpace
