+++ test/test_space.py
@@ -24,6 +24,7 @@ from ridepy.util.spaces_cython import (
     Euclidean2D as CyEuclidean2D,
     Manhattan2D as CyManhattan2D,
     Graph as CyGraph,
+    Grid2D as CyGrid2D,
     Grid2D_QM as CyGrid2D_QM,
 )
 
@@ -164,6 +165,15 @@ def test_CyManhattan2D():
 
 
 def test_CyGrid2D():
+    space = CyGrid2D()
+
+    assert space.d((0, 0), (0, 1)) == 1.0
+    assert space.d((0, 0), (0, 0)) == 0.0
+    assert space.d((0, 0), (1, 1)) == 2
+    assert space.d((0, 0), (0, 0.1)) == 0.1
+
+
+def test_CyGrid2D_QM():
     space = CyGrid2D_QM()
 
     assert space.d((0, 0), (0, 1)) == 1.0
@@ -173,6 +183,22 @@ def test_CyGrid2D():
 
 
 def test_CyGrid2D_velocity():
+    space = CyGrid2D()
+
+    assert space.t((0, 0), (0, 1)) == 1.0
+    assert space.t((0, 0), (0, 0)) == 0.0
+    assert space.t((0, 0), (1, 1)) == 2
+    assert space.t((0, 0), (0, 0.1)) == 0.1
+
+    space = CyGrid2D(velocity=2)
+
+    assert space.t((0, 0), (0, 1)) == 0.5
+    assert space.t((0, 0), (0, 0)) == 0.0
+    assert space.t((0, 0), (1, 1)) == 1
+    assert space.t((0, 0), (0, 0.1)) == 0.05
+
+
+def test_CyGrid2D_QM_velocity():
     space = CyGrid2D_QM()
 
     assert space.t((0, 0), (0, 1)) == 1.0
@@ -188,9 +214,11 @@ def test_CyGrid2D_velocity():
     assert space.t((0, 0), (0, 0.1)) == 0.05
 
 
-def test_CyGrid2D_interpolation():
+@pytest.mark.xfail
+def test_CyGrid2D_QM_interpolation():
 
     def assert_interpolation_equal(a, b):
+        # TODO this should not be necessary anymore, as we want exact values
         assert a[0][0] == pytest.approx(b[0][0])
         assert a[0][1] == pytest.approx(b[0][1])
         assert a[1] == pytest.approx(b[1])
@@ -290,6 +318,16 @@ def test_CyGrid2D_interpolation():
 
 
 def test_CyGrid2D_random_points_discrete():
+    space = CyGrid2D()
+    random_points = [space.random_point() for _ in range(1000)]
+    for random_point in random_points:
+        assert 0 <= random_point[0] <= 10
+        assert 0 <= random_point[1] <= 10
+        assert random_point[0] % 1 == 0
+        assert random_point[1] % 1 == 0
+
+
+def test_CyGrid2D_QM_random_points_discrete():
     space = CyGrid2D_QM()
     random_points = [space.random_point() for _ in range(1000)]
     for random_point in random_points:
