+++ test/test_space.py
@@ -214,6 +214,108 @@ def test_CyGrid2D_QM_velocity():
     assert space.t((0, 0), (0, 0.1)) == 0.05
 
 
+def test_CyGrid2D_interpolation():
+
+    def assert_interpolation_equal(a, b):
+        # TODO this should not be necessary anymore, as we want exact values
+        assert a[0][0] == pytest.approx(b[0][0])
+        assert a[0][1] == pytest.approx(b[0][1])
+        assert a[1] == pytest.approx(b[1])
+
+    space = CyGrid2D()
+
+    assert_interpolation_equal(space.interp_dist((0, 0), (0, 1), 1.0), ((0, 0), 0))
+    # assert space.interp_dist((0, 1), (0, 1), 1.0) == ((1, 0), 0) # this is undefined behavior
+
+    # ========================================
+
+    # X---X   X   X   X
+    #     V
+    # X   X   X   X   X
+    #
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((0, 0), (1, 1), 1.0), ((0, 1), 0))
+
+    # X   X   X   X   X
+    # A
+    # X---X   X   X   X
+    #
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((1, 1), (0, 0), 1.0), ((1, 0), 0))
+
+    # ========================================
+
+    # X---X   X   X   X
+    #     |
+    # X   X   X   X   X
+    #     V
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((0, 0), (2, 1), 0.6), ((2, 1), 0.6))
+    assert_interpolation_equal(space.interp_dist((0, 0), (2, 1), 1.0), ((1, 1), 0))
+    assert_interpolation_equal(space.interp_dist((0, 0), (2, 1), 1.6), ((1, 1), 0.6))
+    assert_interpolation_equal(space.interp_dist((0, 0), (2, 1), 2.6), ((0, 1), 0.6))
+
+    # X   X   X   X   X
+    # A
+    # X   X   X   X   X
+    # |
+    # X---X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((2, 1), (0, 0), 0.6), ((0, 0), 0.6))
+    assert_interpolation_equal(space.interp_dist((2, 1), (0, 0), 1.0), ((1, 0), 0))
+    assert_interpolation_equal(space.interp_dist((2, 1), (0, 0), 1.6), ((1, 0), 0.6))
+    assert_interpolation_equal(space.interp_dist((2, 1), (0, 0), 2.6), ((2, 0), 0.6))
+
+    # ========================================
+
+    # X---X   X   X   X
+    # V
+    # X   X   X   X   X
+    #
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((1, 0), (0, 1), 1.0), ((0, 0), 0))
+
+    # X   X   X   X   X
+    #     A
+    # X---X   X   X   X
+    #
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((0, 1), (1, 0), 1.0), ((1, 1), 0))
+
+    # ========================================
+
+    # X-->X   X   X   X
+    # |
+    # X   X   X   X   X
+    # |
+    # X   X   X   X   X
+
+    assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 0.6), ((0, 1), 0.6))
+    assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 1.0), ((0, 0), 0))
+    assert_interpolation_equal(
[…]
