+++ test/test_space.py
@@ -297,9 +297,7 @@ def test_CyGrid2D_interpolation():
 
     assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 0.6), ((0, 1), 0.6))
     assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 1.0), ((0, 0), 0))
-    assert_interpolation_equal(
-        space.interp_dist((2, 0), (0, 1), 1.6), ((0, 0), 0.6)
-    )  # this is broken (returns ((1, 1), 0.6), which is incompatible with the previous result)
+    assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 1.6), ((0, 0), 0.6))
     assert_interpolation_equal(space.interp_dist((2, 0), (0, 1), 2.6), ((1, 0), 0.6))
 
     # X   X   X   X   X
@@ -310,9 +308,7 @@ def test_CyGrid2D_interpolation():
 
     assert_interpolation_equal(space.interp_dist((0, 1), (2, 0), 0.6), ((2, 0), 0.6))
     assert_interpolation_equal(space.interp_dist((0, 1), (2, 0), 1.0), ((2, 1), 0))
-    assert_interpolation_equal(
-        space.interp_dist((0, 1), (2, 0), 1.6), ((2, 1), 0.6)
-    )  # this is broken
+    assert_interpolation_equal(space.interp_dist((0, 1), (2, 0), 1.6), ((2, 1), 0.6))
     assert_interpolation_equal(space.interp_dist((0, 1), (2, 0), 2.6), ((1, 1), 0.6))
 
 
