+++ src/ridepy/util/spaces_cython/spaces.pyx
@@ -228,6 +228,12 @@ cdef class Grid2D(Manhattan2D):
         self.dm = dm
         self.velocity = velocity
 
+    def d(self, u, v):
+        return abs(u[0] - v[0]) * self.dn + abs(u[1] - v[1]) * self.dm
+
+    def t(self, u, v):
+        return self.d(u, v) / self.velocity
+
     def interp_dist(self, u, v, double dist_to_dest):
         d_r = dist_to_dest # Remaining distance
         d_t = self.d(u, v) # Total distance
@@ -238,19 +244,19 @@ cdef class Grid2D(Manhattan2D):
 
         if d_r == d_t:
             # We are still at the origin
-            print("We are still at the origin")
+            # print("We are still at the origin")
             i = u[0]
             j = u[1]
             d_j = 0
         elif d_r == 0:
             # We have reached the destination
-            print("We have reached the destination")
+            # print("We have reached the destination")
             i = v[0]
             j = v[1]
             d_j = 0
         else:
             # We are under way
-            print("We are under way")
+            # print("We are under way")
             w = (v[0] - u[0], v[1] - u[1])
 
             d_vert = abs(w[0]) * d_n
@@ -258,82 +264,82 @@ cdef class Grid2D(Manhattan2D):
 
             if (w[0] < 0 < w[1]) or (w[0] > 0 > w[1]) or (w[1] == 0):
                 # Going vertically, first
-                print("Going vertically, first")
+                # print("Going vertically, first")
                 if d_e <= d_vert:
                     # We have not made the turn.
-                    print("We have not made the turn")
+                    # print("We have not made the turn")
                     j = u[1]
                     # Determine next node:
                     if w[0] < 0:
                         # Going upwards
-                        print("Going upwards")
+                        # print("Going upwards")
                         d_e_d = d_e
                         k_d_e_d = m.ceil(d_e_d / d_n)
                         i = u[0] - k_d_e_d
                         d_j = k_d_e_d * d_n - d_e_d
                     else:
                         # Going downwards
-                        print("Going downwards")
+                        # print("Going downwards")
                         d_e_d = d_e
                         k_d_e_d = m.ceil(d_e_d / d_n)
                         i = u[0] + k_d_e_d
                         d_j = k_d_e_d * d_n - d_e_d
                 else:
                     # We have made the turn.
-                    print("We have made the turn")
+                    # print("We have made the turn")
                     i = v[0]
                     # Determine next node:
                     if w[1] < 0:
                         # Going left
-                        print("Going left")
+                        # print("Going left")
                         d_e_d = d_e - d_vert
                         k_d_e_d = m.ceil(d_e_d / d_m)
                         j = u[1] - k_d_e_d
                         d_j =  k_d_e_d * d_m - d_e_d
                     else:
                         # Going right
-                        print("Going right")
+                        # print("Going right")
                         d_e_d = d_e - d_vert
                         k_d_e_d = m.ceil(d_e_d / d_m)
                         j = u[1] + k_d_e_d
                         d_j =  k_d_e_d * d_m - d_e_d
             else:
                 # Going horizontally, first
-                print("Going horizontally, first")
+                # print("Going horizontally, first")
                 if d_e <= d_hori:
                     # We have not made the turn
-                    print("We have not made the turn")
+                    # print("We have not made the turn")
                     i = u[0]
                     # Determine the next node:
                     if w[1] < 0:
                         # Going left
-                        print("Going left")
[…]
