### 2024-02-02 994da96
Added Grid2D space (#237)
* added Grid2D space

* added failing interpolation test

* more explanation

* more grid fun

* more grid fun

* more grid fun: tests

* more bugs fixed

* finally works, tests passing

* getting there

* all done for now

* actually done for now

* gitignore
