+++ src/ridepy/util/spaces_cython/spaces.pyx
@@ -4,6 +4,7 @@ import random
 import warnings
 import itertools as it
 import networkx as nx
+import math as m
 
 from libcpp.vector cimport vector
 from libcpp.pair cimport pair
@@ -193,6 +194,277 @@ cdef class Manhattan2D(TransportSpace):
     def __reduce__(self):
         return self.__class__, (self.velocity, self.coord_range)
 
+
+cdef class Grid2D(Manhattan2D):
+    """
+    Fast 2D grid network space, realized as a discrete 2D space with Manhattan metric.
+
+    Note that this space does not enforce the discrete integer coordinate pairs it is
+    supposed to be used with. This means that you have to take care yourself to only
+    introduce integer coordinate pairs. This implies that the distance and time functions
+    ``Grid2D.d`` and ``Grid2D.t`` will in fact as if the space were continuous.
+    The discrete/graph characteristic of the space is encoded in the interpolation and
+    random point generation functions.
+
+    The interpolation function will always assume the shortest path with no turns or a
+    single right turn.
+    """
+    def __cinit__(self, int n=10, int m=10, double dn=1, double dm=1, double velocity=1):
+        self.loc_type = LocType.R2LOC
+        self.derived_ptr = self.u_space.space_r2loc_ptr = new CManhattan2D(velocity)
+
+    def __init__(self, n=10, m=10, dn=1, dm=1, velocity=1):
+        """
+        Parameters
+        ----------
+        velocity
+            constant velocity to compute travel time, optional. default: 1
+        """
+        TransportSpace.__init__(self, loc_type=LocType.R2LOC)
+
+        self.n = n
+        self.m = m
+        self.dn = dn
+        self.dm = dm
+        self.velocity = velocity
+
+    def d(self, u, v):
+        return abs(u[0] - v[0]) * self.dn + abs(u[1] - v[1]) * self.dm
+
+    def t(self, u, v):
+        return self.d(u, v) / self.velocity
+
+    def interp_dist(self, u, v, double dist_to_dest):
+        d_r = dist_to_dest # Remaining distance
+        d_t = self.d(u, v) # Total distance
+        d_e = d_t - d_r # Elapsed distance
+        d_n = self.dn # Vertical grid spacing
+        d_m = self.dm # Horizontal grid spacing
+
+
+        if d_r == d_t:
+            # We are still at the origin
+            # print("We are still at the origin")
+            i = u[0]
+            j = u[1]
+            d_j = 0
+        elif d_r == 0:
+            # We have reached the destination
+            # print("We have reached the destination")
+            i = v[0]
+            j = v[1]
+            d_j = 0
+        else:
+            # We are under way
+            # print("We are under way")
+            w = (v[0] - u[0], v[1] - u[1])
+
+            d_vert = abs(w[0]) * d_n
+            d_hori = abs(w[1]) * d_m
+
+            if (w[0] < 0 < w[1]) or (w[0] > 0 > w[1]) or (w[1] == 0):
+                # Going vertically, first
+                # print("Going vertically, first")
+                if d_e <= d_vert:
+                    # We have not made the turn.
+                    # print("We have not made the turn")
+                    j = u[1]
+                    # Determine next node:
+                    if w[0] < 0:
+                        # Going upwards
+                        # print("Going upwards")
+                        d_e_d = d_e
+                        k_d_e_d = m.ceil(d_e_d / d_n)
+                        i = u[0] - k_d_e_d
+                        d_j = k_d_e_d * d_n - d_e_d
+                    else:
+                        # Going downwards
+                        # print("Going downwards")
+                        d_e_d = d_e
+                        k_d_e_d = m.ceil(d_e_d / d_n)
+                        i = u[0] + k_d_e_d
[…]
