+++ doc/notebooks/introduction_simulation_set.py
@@ -6,11 +6,11 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.14.5
+#       jupytext_version: 1.15.2
 #   kernelspec:
-#     display_name: Python (ridepy39)
+#     display_name: Python 3.9 (ridepy)
 #     language: python
-#     name: ridepy39
+#     name: ridepy
 # ---
 
 # # RidePy Introduction: Multiple Simulations
