+++ doc/conf.py
@@ -18,7 +18,7 @@ sys.path.insert(0, os.path.abspath("../"))
 # -- Project information -----------------------------------------------------
 
 project = "ridepy"
-copyright = "2023, Felix Jung and Debsankha Manik"
+copyright = "2024, Felix Jung and Debsankha Manik"
 author = "Felix Jung and Debsankha Manik"
 
 # Aesthetics
@@ -45,6 +45,8 @@ extensions = [
     "sphinx_rtd_theme",
     "sphinxcontrib.napoleon",  # for sane autodoc
     "alabaster",  # theme
+    "myst_nb",  # for jupyter notebook support
+    "nbsphinx_link",
 ]
 
 # Napoleon settings
@@ -77,7 +79,6 @@ autodoc_type_aliases = {
 }
 autodoc_inherit_docstrings = False
 
-
 # Intersphinx mapping
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
@@ -94,7 +95,13 @@ templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The suffix of source filenames.
-source_suffix = ".rst"
+source_suffix = {
+    ".rst": "restructuredtext",
+    ".ipynb": "myst-nb",
+    ".myst": "myst-nb",
+    ".md": "myst-nb",
+}
+
 
 # The encoding of source files.
 # source_encoding = 'utf-8-sig'
@@ -113,3 +120,7 @@ html_theme = "sphinx_rtd_theme"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+
+nb_custom_formats = {
+    ".md": ["jupytext.reads", {"fmt": "mystnb"}],
+}
