### 2024-02-05 b3fb864
should probably replace quickstart with a notebook
