+++ doc/notebooks/introduction.md
@@ -1,19 +1,19 @@
 ---
 jupytext:
   encoding: '# -*- coding: utf-8 -*-'
-  formats: ipynb,py:light,md:myst
+  formats: md:myst
   text_representation:
     extension: .md
     format_name: myst
     format_version: 0.13
-    jupytext_version: 1.14.7
+    jupytext_version: 1.15.2
 kernelspec:
   display_name: Python 3.9 (ridepy)
   language: python
   name: ridepy
 ---
 
-# RidePy Introduction: Basics
+# RidePy Tutorial: Basic simulations
 
 ```{code-cell}
 %matplotlib inline
@@ -58,15 +58,31 @@ pd.set_option("display.width", 1000)
 evf = lambda S, f, **arg: (S, f(S, **arg))
 ```
 
-## Configure the simulation and supply initial values
+## Configuring the simulation and supplying initial values
+### Choosing a simulation space
+
++++
+
+The first important step for performing a RidePy simulation is the choice of the physical space that the simulations should be run on. For this example, we choose the Euclidean 2D space from the `util` package.
 
 ```{code-cell}
-n_buses = 50
-
-initial_location = (0, 0)
-
 space = Euclidean2D()
+```
 
+### Choosing a way of generating requests for transportation
+
+The basis for RidePy simulations are `TransportationRequest`s. Each of these consists of:
+
+- `origin`
+- `destination`
+- `pickup_timewindow_min`
+- `pickup_timewindow_max`
+- `delivery_timewindow_min`
+- `delivery_timewindow_max`
+
+To generate these `TransportationRequest`s, we will use the `RandomRequestGenerator` from the `util` package:
+
+```{code-cell}
 rg = RandomRequestGenerator(
     rate=10,
     max_pickup_delay=3,
@@ -74,12 +90,18 @@ rg = RandomRequestGenerator(
     space=space,
     seed=42,
 )
-
-# create iterator yielding 100 random requests
-transportation_requests = it.islice(rg, 100)
 ```
 
-### Initialize a `FleetState`
+### Setting up a fleet of vehicles
+
+RidePy manages a fleet of vehicles using a `FleetState` object. It consumes a dictionary of `initial_locations` which maps arbitrary vehicle ids to their starting position in the simulation. The number of vehicles to set up is inferred from the number of entries in the `initial_conditions` dict. 
+
+In addition, the fleet state needs to know about the space used for the simulation, and about the desired `dispatcher`. The dispatcher function contains the algorithm that determines how stops to serve incoming requests are scheduled. In this case, we use the included `BruteForceTotalTravelTimeMinimizingDispatcher`.
+
+```{code-cell}
+n_buses = 50
+initial_location = (0, 0)
+```
 
 ```{code-cell}
 fs = SlowSimpleFleetState(
@@ -91,41 +113,78 @@ fs = SlowSimpleFleetState(
 )
 ```
 
-### Perform the simulation
+## Performing a simulation
+
+To run the simulation we first take a slice of, in this case, 100 random requests out of the request generator we set up above...
+
+```{code-cell}
+transportation_requests = it.islice(rg, 100)
+```
+
+...and feed them into `FleetState.simulate`. Note that both of these operations use iterators and no computation actually happens until the iterators are exhausted. For `FleetState.simulate`, this happens when we cast its output into a Python list. In the case of the request generator, the output is an iterator of `TransportationRequest` objects, in the case of the `simulate` method an iterator of `Event` objects. These events describe e.g. that a request was accepted or that a "customer" or "rider" (represented by its `TransportationRequest`) was picked up or delivered to her destination.
 
 ```{code-cell}
[…]
