+++ doc/notebooks/introduction.md
@@ -13,9 +13,16 @@ kernelspec:
   name: ridepy
 ---
 
++++ {"editable": true, "slideshow": {"slide_type": ""}}
+
 # RidePy Tutorial: Basic simulations
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 %matplotlib inline
 
 import itertools as it
@@ -26,6 +33,11 @@ import matplotlib.pyplot as plt
 ```
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 from ridepy.fleet_state import SlowSimpleFleetState
 from ridepy.vehicle_state import VehicleState
 
@@ -39,6 +51,11 @@ from ridepy.util.analytics.plotting import plot_occupancy_hist
 ```
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 # assume dark background for plots?
 dark = False
 
@@ -51,6 +68,11 @@ if dark:
 ```
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 pd.set_option("display.max_rows", 500)
 pd.set_option("display.max_columns", 500)
 pd.set_option("display.width", 1000)
@@ -58,16 +80,25 @@ pd.set_option("display.width", 1000)
 evf = lambda S, f, **arg: (S, f(S, **arg))
 ```
 
++++ {"editable": true, "slideshow": {"slide_type": ""}}
+
 ## Configuring the simulation and supplying initial values
 ### Choosing a simulation space
 
-+++
++++ {"editable": true, "slideshow": {"slide_type": ""}}
 
 The first important step for performing a RidePy simulation is the choice of the physical space that the simulations should be run on. For this example, we choose the Euclidean 2D space from the `util` package.
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 space = Euclidean2D()
 ```
+
++++ {"editable": true, "slideshow": {"slide_type": ""}}
 
 ### Choosing a way of generating requests for transportation
 
@@ -83,6 +114,11 @@ The basis for RidePy simulations are `TransportationRequest`s. Each of these con
 To generate these `TransportationRequest`s, we will use the `RandomRequestGenerator` from the `util` package:
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 rg = RandomRequestGenerator(
     rate=10,
     max_pickup_delay=3,
@@ -92,6 +128,8 @@ rg = RandomRequestGenerator(
 )
 ```
 
++++ {"editable": true, "slideshow": {"slide_type": ""}}
+
 ### Setting up a fleet of vehicles
 
 RidePy manages a fleet of vehicles using a `FleetState` object. It consumes a dictionary of `initial_locations` which maps arbitrary vehicle ids to their starting position in the simulation. The number of vehicles to set up is inferred from the number of entries in the `initial_conditions` dict. 
@@ -99,11 +137,21 @@ RidePy manages a fleet of vehicles using a `FleetState` object. It consumes a di
 In addition, the fleet state needs to know about the space used for the simulation, and about the desired `dispatcher`. The dispatcher function contains the algorithm that determines how stops to serve incoming requests are scheduled. In this case, we use the included `BruteForceTotalTravelTimeMinimizingDispatcher`.
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 n_buses = 50
 initial_location = (0, 0)
 ```
 
 ```{code-cell}
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
[…]
