+++ doc/overview.rst
@@ -15,204 +15,8 @@ The setting
 
 Rather than explicitly modeling the behavior of individuals (classic agent-based simulations), RidePy instead starts with a set of *transportation requests* (denoting the desire of an individual to be transported from A to B within a specified time window), and a *dispatcher* (an algorithm that determines which vehicle should service which requests and in which order). Then it simply simulates these requests arriving in the system, being picked up, and being delivered. Requests that cannot be delivered within the specified time windows are *rejected*.
 
-RidePy makes it easy to experiment with different dispatching algorithms, spatio-temporal request densities, fleet sizes and transport spaces (2-D plane, different graphs). It comes with an `analytics` module that computes from the simulation output various metrics like the statistical distributions of occupancies, waiting times, and detours.
+RidePy makes it easy to experiment with different dispatching algorithms, spatio-temporal request densities, fleet sizes and transport spaces (2D plane, different graphs). It comes with an `analytics` package that from the simulation computes various output metrics like e.g. the statistical distributions of occupancies, waiting times, and detours.
 
-Since the ability to simulate many *requests* and many *vehicles* is important for any quantitative study, we include two powerful ways of speeding up the simulation:
+Since the ability to simulate many *requests* and many *vehicles* is important for any quantitative study RidePy uses `Cython <https://cython.readthedocs.io/en/latest/>`_ to achieve very fast simulations. We provide, out of the box, the ability to choose either pure pythonic or cythonic data structures and algorithms for running the simulations. This way, dispatchers can be defined in the C++ language. See the :ref:`cython overview <using_cython>` for more detail.
 
-Using parallelism:
-   The dispatcher interface prescribes computing a *cost* for serving a request with a
-   certain vehicle. Then the vehicle with the minimum cost is chosen. Since this is an
-   "embarassingly parallelizable" operation, RidePy provides two parallel ways of
-   computing it, out of the box:
-
-   - ``multiprocessing``,
-   - ``OpenMPI``.
-
-Using `cython <https://cython.readthedocs.io/en/latest/>`_:
-   We provide, out of the box, the ability to choose either pure pythonic or cythonic
-   data structures and algorithms for running the simulations. This way, dispatchers can
-   be defined in the C++ language. See :ref:`using_cython` for details.
-
-
-Quickstart
-----------
-
-In this section, we will demonstrate how to run a simple simulation.
-
-Note that several introductory notebooks for easy experimentation are available under ``doc/notebooks`` in the git repository, see also :ref:`first steps <first_steps>`. Alternatively you can directly have a look at them :doc:`here <notebooks>`.
-
-Generate requests
-^^^^^^^^^^^^^^^^^
-
-First we need to generate a sequence of :class:`TransportationRequest
-<data_structures.TransportationRequest>`. Each ``TransportationRequest`` consists of:
-
-  - ``origin``,
-  - ``destination``,
-  - ``pickup_timewindow_min``,
-  - ``pickup_timewindow_max``,
-  - ``delivery_timewindow_min``,
-  - ``delivery_timewindow_max``. 
-
-We will use the :mod:`request_generators <util.request_generators>` module to
-generate some requests with random origins and destinations:
-
[…]
