### 2024-02-05 ccf5a34
notebooks done for now
