+++ doc/notebooks/introduction.md
@@ -15,70 +15,8 @@ kernelspec:
 
 +++ {"editable": true, "slideshow": {"slide_type": ""}}
 
-# RidePy Tutorial: Basic simulations
-
-```{code-cell}
----
-editable: true
-slideshow:
-  slide_type: ''
----
-%matplotlib inline
-
-import itertools as it
-import math as m
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-```
-
-```{code-cell}
----
-editable: true
-slideshow:
-  slide_type: ''
----
-from ridepy.fleet_state import SlowSimpleFleetState
-from ridepy.vehicle_state import VehicleState
-
-from ridepy.util.dispatchers import BruteForceTotalTravelTimeMinimizingDispatcher
-
-from ridepy.util.request_generators import RandomRequestGenerator
-from ridepy.util.spaces import Euclidean2D
-
-from ridepy.util.analytics import get_stops_and_requests
-from ridepy.util.analytics.plotting import plot_occupancy_hist
-```
-
-```{code-cell}
----
-editable: true
-slideshow:
-  slide_type: ''
----
-# assume dark background for plots?
-dark = False
-
-if dark:
-    default_cycler = plt.rcParams["axes.prop_cycle"]
-    plt.style.use("dark_background")
-    plt.rcParams["axes.prop_cycle"] = default_cycler
-    plt.rcParams["axes.facecolor"] = (1, 1, 1, 0)
-    plt.rcParams["figure.facecolor"] = (1, 1, 1, 0)
-```
-
-```{code-cell}
----
-editable: true
-slideshow:
-  slide_type: ''
----
-pd.set_option("display.max_rows", 500)
-pd.set_option("display.max_columns", 500)
-pd.set_option("display.width", 1000)
-
-evf = lambda S, f, **arg: (S, f(S, **arg))
-```
+# RidePy Tutorial 1: Basics
+This tutorial covers the basics of using RidePy: Setting up a simulation, running it, and doing basic analytics on it.
 
 +++ {"editable": true, "slideshow": {"slide_type": ""}}
 
@@ -87,14 +25,16 @@ evf = lambda S, f, **arg: (S, f(S, **arg))
 
 +++ {"editable": true, "slideshow": {"slide_type": ""}}
 
-The first important step for performing a RidePy simulation is the choice of the physical space that the simulations should be run on. For this example, we choose the Euclidean 2D space from the `util` package.
+The first important step for performing a RidePy simulation is the choice of the physical space that the simulations should be run on. For this example, we choose the Euclidean 2D `TransportSpace` from the `util` package.
 
-```{code-cell}
+```{code-cell} ipython3
 ---
 editable: true
 slideshow:
   slide_type: ''
 ---
+from ridepy.util.spaces import Euclidean2D
+
 space = Euclidean2D()
 ```
 
@@ -113,12 +53,14 @@ The basis for RidePy simulations are `TransportationRequest`s. Each of these con
 
 To generate these `TransportationRequest`s, we will use the `RandomRequestGenerator` from the `util` package:
 
-```{code-cell}
+```{code-cell} ipython3
 ---
 editable: true
 slideshow:
   slide_type: ''
 ---
+from ridepy.util.request_generators import RandomRequestGenerator
+
 rg = RandomRequestGenerator(
     rate=10,
     max_pickup_delay=3,
@@ -134,9 +76,9 @@ rg = RandomRequestGenerator(
 
 RidePy manages a fleet of vehicles using a `FleetState` object. It consumes a dictionary of `initial_locations` which maps arbitrary vehicle ids to their starting position in the simulation. The number of vehicles to set up is inferred from the number of entries in the `initial_conditions` dict. 
 
[…]
