+++ doc/notebooks/introduction_cython.md
@@ -6,69 +6,52 @@ jupytext:
     extension: .md
     format_name: myst
     format_version: 0.13
-    jupytext_version: 1.14.7
+    jupytext_version: 1.15.2
 kernelspec:
   display_name: Python 3.9 (ridepy)
   language: python
   name: ridepy
 ---
 
-# RidePy Introduction: Simulations with Cython
+# RidePy Tutorial 2: Faster simulations using Cython
 
-```{code-cell}
-%matplotlib inline
++++
+
+In Tutorial 1 we have seen how to set up a basic simulation, run, and analyze it. This tutorial is concerned with significantly increasing simulation performance by replacing the crucial components with their Cython equivalents. 
+
+To demonstrate this, we will first show the complete process again in a condensed form, using the Python components. Afterwards, we will do the same using the Cython components.
+
+The result processing and data analytics steps will not be covered again, as they are identical to using the Python components.
+
++++
+
+## Simulation using Python components
+
+First, we import all necessary **Python** components:
+
+```{code-cell} ipython3
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
+from ridepy.util.spaces import Euclidean2D
+from ridepy.util.request_generators import RandomRequestGenerator
+from ridepy.fleet_state import SlowSimpleFleetState
+from ridepy.vehicle_state import VehicleState
+from ridepy.util.dispatchers import BruteForceTotalTravelTimeMinimizingDispatcher
 
 import itertools as it
-import math as m
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
 ```
 
-```{code-cell}
-from ridepy.fleet_state import SlowSimpleFleetState
-from ridepy.vehicle_state_cython import VehicleState
+Now we can set up the simulation as explained in Tutorial 1:
 
-from ridepy.util.dispatchers_cython import (
-    BruteForceTotalTravelTimeMinimizingDispatcher,
-    SimpleEllipseDispatcher,
-)
-
-from ridepy.util.request_generators import RandomRequestGenerator
-from ridepy.util.spaces_cython import Euclidean2D
-from ridepy.data_structures_cython import TransportationRequest
-
-from ridepy.util.analytics import get_stops_and_requests
-from ridepy.util.analytics.plotting import plot_occupancy_hist
-```
-
-```{code-cell}
-# assume dark background for plots?
-dark = False
-
-if dark:
-    default_cycler = plt.rcParams["axes.prop_cycle"]
-    plt.style.use("dark_background")
-    plt.rcParams["axes.prop_cycle"] = default_cycler
-    plt.rcParams["axes.facecolor"] = (1, 1, 1, 0)
-    plt.rcParams["figure.facecolor"] = (1, 1, 1, 0)
-```
-
-```{code-cell}
-pd.set_option("display.max_rows", 500)
-pd.set_option("display.max_columns", 500)
-pd.set_option("display.width", 1000)
-
-evf = lambda S, f, **arg: (S, f(S, **arg))
-```
-
-## Configure the simulation and supply initial values
-
-```{code-cell}
-n_buses = 50
-
-initial_location = (0, 0)
-
+```{code-cell} ipython3
+---
+editable: true
+slideshow:
+  slide_type: ''
+---
 space = Euclidean2D()
 
 rg = RandomRequestGenerator(
@@ -76,60 +59,109 @@ rg = RandomRequestGenerator(
     max_pickup_delay=3,
     max_delivery_delay_rel=1.9,
     space=space,
-    request_class=TransportationRequest,
     seed=42,
 )
 
-# create iterator yielding 100 random requests
-transportation_requests = it.islice(rg, 1000)
-```
+n_buses = 50
[…]
