+++ doc/setup.rst
@@ -1,0 +1,61 @@
+Installation
+============
+
+This section explains how to set up RidePy on your machine.
+
+Prerequisites
+-------------
+
+-  Python 3.9
+
+You should probably use an environment. For example, using
+`conda <https://www.anaconda.com/>`__:
+
+.. code:: sh
+
+    conda create -n ridepy python=3.9
+    conda activate ridepy
+
+
+Finally, a C++ build environment and the `Boost C++ Libraries <https://www.boost.org/>`__
+are necessary if you want or need to build the Cython/C++ part from source. This step
+can be skipped when installing the Python Wheel via ``pip`` on supported platforms
+(currently only x86-64 Linux).
+
+On Debian-based Linux distributions, these dependencies may be installed as follows:
+
+.. code:: sh
+
+    sudo apt-get update && sudo apt-get -y install libboost-all-dev build-essential
+
+User Installation
+-----------------
+
+Just run
+
+.. code:: sh
+
+    pip install ridepy
+
+If you prefer, you can also use clone the git repository instead:
+
+.. code:: sh
+
+    git clone --recurse-submodules https://github.com/PhysicsOfMobility/ridepy.git
+    cd ridepy
+    pip install -e .
+
+
+Developer Installation
+----------------------
+
+.. code:: sh
+
+    git clone --recurse-submodules https://github.com/PhysicsOfMobility/ridepy.git
+    cd ridepy
+    pip install -e ".[dev,doc]"
+    make -C doc html
+    pre-commit install
+    pytest
+
+The built documentation can be found in ``doc/_build/html/index.html``.
