+++ doc/support.rst
@@ -1,0 +1,3 @@
+Obtaining support
+=================
+
