+++ doc/issues.rst
@@ -1,0 +1,3 @@
+Reporting issues
+================
+
