+++ doc/reference.rst
@@ -1,5 +1,5 @@
-Reference
-=========
+API Reference
+=============
 
 .. toctree::
    :maxdepth: 5
