+++ doc/issues.rst
@@ -1,3 +1,5 @@
 Reporting issues
 ================
 
+Should you encounter any problems when using RidePy or have a feature request, please don't hesitate to `submit an issue <https://github.com/PhysicsOfMobility/ridepy/issues/new>`__.
+
