+++ doc/index.rst
@@ -1,12 +1,5 @@
-.. ridepy documentation master file, created by
-   sphinx-quickstart on Sun Mar 21 12:40:01 2021.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 Welcome to ridepy's documentation!
 ========================================
-.. First include the project README
-.. include:: ../README.rst
 
 .. toctree::
    :maxdepth: 5
