+++ doc/index.rst
@@ -1,14 +1,24 @@
-Welcome to ridepy's documentation!
-========================================
+Welcome to RidePy's documentation!
+===================================
+
+RidePy is a scientific Python library for simulating modern on-demand transit systems such as ridepooling.
+
+To better understand what RidePy is all about probably should start by reading the :doc:`overview section <overview>`. For instructions on installation, contributing and support, head over to the :doc:`*How to...* section <howto>`. And to dive right in and get started with using RidePy for a simple simulation, try the :doc:`tutorial <notebooks>`.
+
+In addition to these general guides, we offer a commented :doc:`API reference <reference>`, in which we explain the building blocks and concepts that RidePy is based on in more detail, and how you can leverage them for your specific application. This section is far more technical than the hands-on tutorial, but allows for a deeper understanding.
+
+Finally, we now have a helpful (currently still incomplete) :doc:`glossary <glossary>`, which details both general terms and concepts as well as those specific to the analytics module.
+
+For quick access to the documentation of specific parts of the API, refer to the hierarchically structured :ref:`Python module index <modindex>`. There is also an alphabetical :ref:`documentation index <genindex>`.
 
 .. toctree::
-   :maxdepth: 5
-   :hidden:
+    :maxdepth: 2
+    :hidden:
 
-   overview
-   howto
-   notebooks
-   glossary
-   reference
-   Python module index <modindex>
-   genindex
+    overview
+    howto
+    notebooks
+    glossary
+    reference
+    Python module index <modindex>
+    genindex
