### 2024-02-07 5cd6899
updated installation doc
