+++ doc/setup.rst
@@ -1,35 +1,44 @@
-Installation
-============
+.. highlight:: bash
 
-This section explains how to set up RidePy on your machine.
+Setting up RidePy
+=================
+
+This section explains how to set up RidePy on your machine in various ways.
 
 Prerequisites
 -------------
 
--  Python 3.9
+Currently, using Python 3.9 is your best option.
 
-You should probably use an environment. For example, using
-`conda <https://www.anaconda.com/>`__:
+In addition, you should probably use a Python environment for keeping things clean. The following commands should work for Linux and macOS. On Windows, things might have to be done marginally different.
+
+We recommend using either `Anaconda <https://www.anaconda.com/>`__ (which has the added benefit of being able to handle different Python versions):
 
 .. code:: sh
 
     conda create -n ridepy python=3.9
     conda activate ridepy
 
-
-Finally, a C++ build environment and the `Boost C++ Libraries <https://www.boost.org/>`__
-are necessary if you want or need to build the Cython/C++ part from source. This step
-can be skipped when installing the Python Wheel via ``pip`` on supported platforms
-(currently only x86-64 Linux).
-
-On Debian-based Linux distributions, these dependencies may be installed as follows:
+or simply ``venv`` from the Python standard library (assuming you are already on Python 3.9, alternatively you can additionally use `pyenv <https://github.com/pyenv/pyenv>`__ for managing Python versions):
 
 .. code:: sh
 
-    sudo apt-get update && sudo apt-get -y install libboost-all-dev build-essential
+    python -m venv <venv directory path of choice>
+    source <venv directory path of choice>/bin/activate
+
+Finally, it may be necessary that you build the Cython/C++ part from source. If you are on a supported platform (currently only x86-64 Linux) and install RidePy via ``pip`` from the PyPI repository, this step can be skipped. Otherwise, you will need a C++ build environment and the `Boost C++ Libraries <https://www.boost.org/>`__.
+
+On Debian-based Linux distributions, both of these can be installed as follows:
+
+.. code:: sh
+
+    sudo apt-get update && sudo apt-get -y install build-essential libboost-all-dev
+
+Installation
+------------
 
 User Installation
------------------
+~~~~~~~~~~~~~~~~~
 
 Just run
 
@@ -37,7 +46,7 @@ Just run
 
     pip install ridepy
 
-If you prefer, you can also use clone the git repository instead:
+If you prefer, you can also clone the git repository instead (for which you need ``git``):
 
 .. code:: sh
 
@@ -47,15 +56,67 @@ If you prefer, you can also use clone the git repository instead:
 
 
 Developer Installation
-----------------------
+~~~~~~~~~~~~~~~~~~~~~~
+
+Note that the editable install (``-e`` flag) does only serve its purpose when editing the Python components. For changes to the Cython/C++ components to come into effect, the ``pip install`` command has to be executed again to build the Cython/C++.
 
 .. code:: sh
 
     git clone --recurse-submodules https://github.com/PhysicsOfMobility/ridepy.git
     cd ridepy
     pip install -e ".[dev,doc]"
-    make -C doc html
     pre-commit install
+
+To build the documentation to ``doc/_build/html``, additionally execute the following command:
+
+.. code:: sh
+
+    sphinx-build -j3 -b html doc doc/_build/html
+
+
+Testing the installation
+------------------------
+
+To check whether the installation was successful, you may run the automated test suite, which should report no errors:
+
+.. code:: sh
+
     pytest
 
 The built documentation can be found in ``doc/_build/html/index.html``.
+
+Updating the installation
+-------------------------
+
+User installation
+~~~~~~~~~~~~~~~~~
+
[…]
