### 2024-02-07 c6fdea4
installation details
