+++ doc/issues.rst
@@ -1,5 +1,18 @@
-Reporting issues
-================
+Reporting an issue
+==================
 
 Should you encounter any problems when using RidePy or have a feature request, please don't hesitate to `submit an issue <https://github.com/PhysicsOfMobility/ridepy/issues/new>`__.
 
+First, check whether the issue has been reported already. If this is the case, feel free to contribute to the discussion, or, if you're inclined, even try to fix it yourself. Pull requests are always welcome, head over to the documentation for :doc:`contributors <contributing>` to learn more.
+
+If you in fact have run into an unreported issue, please describe clearly which steps have lead to the problem and, if possible, provide a minimum working example. GitHub issues support markdown syntax which makes it easy to format them in a concise way. Use code cells for highlighted code like this:
+
+.. code::markdown
+
+    ```python
+
+    def my_broken_function(a: int, b:float):
+        result = do_stuff()
+        return 42 + result
+
+    ```
