+++ doc/support.rst
@@ -1,3 +1,7 @@
 Obtaining support
 =================
 
+Hopefully the most helpful resource in learning how to use RidePy and dealing with problems will be this documentation, along with the :doc:`tutorial <notebooks>`. However, if you fail to understand parts of the documentation, find them to be confusing, misleading, or missing important details, please submit an :doc:`issue <issues>` on GitHub so that we can try to fix it.
+
+In case you need more interactive support, aside from asking friends or coworkers, your best option is to either also submit a GitHub :doc:`issue <issues>` detailing you question, or consult the authors directly for help via Email.
+
