### 2024-02-07 7612102
removed unused sphinx extension
