### 2024-02-07 275019a
kernel names removed from notebooks
