+++ doc/notebooks/introduction_cython.md
@@ -8,9 +8,7 @@ jupytext:
     format_version: 0.13
     jupytext_version: 1.15.2
 kernelspec:
-  display_name: Python 3.9 (ridepy)
   language: python
-  name: ridepy
 ---
 
 # RidePy Tutorial 2: Faster simulations using Cython
