+++ doc/notebooks/introduction_simulation_set.md
@@ -7,9 +7,7 @@ jupytext:
     format_version: 0.13
     jupytext_version: 1.15.2
 kernelspec:
-  display_name: Python 3.9 (ridepy)
   language: python
-  name: ridepy
 ---
 
 # RidePy Tutorial 4: Parallelization
