### 2024-02-07 957fe64
kernelspec removed entirely from notebooks
