+++ doc/notebooks/introduction_cython_graph.md
@@ -6,8 +6,6 @@ jupytext:
     format_name: myst
     format_version: 0.13
     jupytext_version: 1.15.2
-kernelspec:
-  language: python
 ---
 
 # RidePy Tutorial 3: Simulations on Graphs
