+++ doc/notebooks/introduction.md
@@ -7,8 +7,6 @@ jupytext:
     format_name: myst
     format_version: 0.13
     jupytext_version: 1.15.2
-kernelspec:
-  language: python
 ---
 
 +++ {"editable": true, "slideshow": {"slide_type": ""}}
