### 2024-02-07 4471b47
238 improve documentation (#240)
* moved notebooks to doc

* moved myst markdown for notebooks

* notebooks are now rendered

* should probably replace quickstart with a notebook

* update readme

* nicified notebooks section

* updated intro notebook

* updated intro notebook

* more doc refactoring

* more doc refactoring

* notebooks done for now

* less clutter

* more restructuring

* removed readme from the doc

* removed readme from the doc

* overview

* cython updates

* updated installation doc

* installation, supress warnings

* installation details

* installation details

* howto
[…]
