+++ test/test_integration.py
@@ -15,22 +15,26 @@ def test_notebooks(tmp_path):
     tmp_py_path = tmp_path / "notebook.py"
     tmp_ipy_path = tmp_path / "notebook.ipy"
 
-    for fpath in Path("notebooks").glob("*.py"):
+    for fpath in Path("doc/notebooks").glob("*.md"):
         print(fpath)
 
         # This stunt is necessary to deal with magic commands
+
         # Constraints:
-        # - Magics must not be in the py light script files, because black will complain.
-        #   Also jupytext doesn't do this on default. This can be changed by modifying the
+        #
+        # - Magics must not be in py light script files, because black will complain.
+        #   Also, jupytext doesn't do this on default. This can be changed by modifying the
         #   notebooks' metadata, but this is obviously not practical to do for every new notebook.
-        # - Magics must be there for the test to work, as e.g. `%time result = my_expensive_computation()`
+        # - Magics must be there for the test to work, as e.g. ``%time result = my_expensive_computation()``
         #   will leave `result` undefined if the line is commented by jupytext
-        # - ipython will understand magics, but only accept them in files ending with `.ipy`
-        # - jupytext will not store anything to files ending with `.ipy`
-        # - we only want to keep `.py` files in the notebooks directory (apart from `.ipynb` files, locally)
-        # this leaves us with this garbage.
+        # - ipython will understand magics, but only accept them in files ending with ``.ipy``
+        # - jupytext will not store anything to files ending with ``.ipy``
+        # - we only want to keep `.md` files in MyST md syntax in the ``doc/notebooks`` directory
+        #
+        # This leaves us with this garbage.
+
         jupytext.write(
-            jupytext.read(fpath, fmt="py"),
+            jupytext.read(fpath, fmt="md:myst"),
             tmp_py_path,
             fmt={"extension": "py", "comment_magics": False},
         )
