+++ doc/howto.rst
@@ -1,0 +1,11 @@
+How to...
+=========
+
+.. toctree::
+   :maxdepth: 1
+
+    Install RidePy? <setup>
+    Contribute? <contributing>
+    Report an issue? <issues>
+    Obtain support? <support>
+
