### 2024-02-07 #6241 @fxjung
@xoolive Thanks a lot for pointing out the documentation system guide! This is a very helpful reference, indeed. I've tried to restructure and improve documentation over the course of the last days, the current version is up on [ridepy.org](https://ridepy.org). It doesn't fully adapt the scheme (yet), technical reference and explanations are still a bit intertwined, and there is only one technical how-to guide.

I have integrated the jupyter notebooks in the documentation, that's actually a great suggestion. Also, I have refactored/rewritten substantial parts of them.

I hope, that the library and its documentation are much more accessible and understandable now.
