### 2024-02-08 1cf0645
moved lru cache module, added analytics modules to sphinx
