### 2024-02-09 3ad80ae
added system stoplist length computation
