+++ src/ridepy/util/analytics/system.py
@@ -1,4 +1,4 @@
-from typing import Optional, Any, Dict
+from typing import Optional, Any, Dict, Union
 
 import pandas as pd
 
@@ -7,7 +7,7 @@ def get_system_quantities(
     stops: pd.DataFrame,
     requests: pd.DataFrame,
     params: Optional[dict[str, dict[str, Any]]] = None,
-) -> Dict[str, pd.DataFrame]:
+) -> dict[str, Union[int, float]]:
     """
     Compute various quantities aggregated for the entire simulation.
 
@@ -35,6 +35,9 @@ def get_system_quantities(
     - **avg_waiting_time**
     - **rejection_ratio**
     - **median_stoplist_length** -- median per-vehicle stoplist length, taken over all "stoplist states" (vehicles x time)
+    - **median_stoplist_length** -- median per-vehicle stoplist length, taken over all "stoplist states" (vehicles x time)
+    - **mean_system_stoplist_length** -- arithmetic mean system-wide stoplist length, taken over all "stoplist states" (time)
+    - **mean_system_stoplist_length** -- arithmetic mean system-wide stoplist length, taken over all "stoplist states" (time)
     - **avg_detour**
     - **(avg_system_stoplist_length_service_time)**
     - **(avg_system_stoplist_length_submission_time)**
@@ -52,18 +55,9 @@ def get_system_quantities(
 
     Returns
     -------
-    dict containing the aforementioned observables
+    system_quantities
+        dict containing the aforementioned observables
 
-    Parameters
-    ----------
-    stops
-        Stops dataframe
-    requests
-        Requests dataframe
-
-    Returns
-    -------
-    dict containing the aforementioned observables
     """
     serviced_requests = (
         requests[requests[("rejected", "timestamp")].isna()]
@@ -122,7 +116,7 @@ def get_system_quantities(
     # submission events, pickup stops, delivery stops and internal stops
     # (initial and final stops). State changes at the same timestamp are merged.
     # It assumes that the stoplist is ordered by timestamp.
-    median_stoplist_length = (
+    stoplist_length_ensemble = (
         event_log["event_type"]
         .map(dict(submission=2, pickup=-1, dropoff=-1))
         .fillna(0)  # internal stops remain, incur no stoplist length delta
@@ -130,8 +124,24 @@ def get_system_quantities(
         .sum()  # merge state changes at same time
         .groupby("vehicle_id")
         .cumsum()  # integrate over stoplist length changes for each vehicle
-        .median()
     )
+
+    median_stoplist_length = stoplist_length_ensemble.median()
+    mean_stoplist_length = stoplist_length_ensemble.mean()
+
+    system_stoplist_length_ensemble = (
+        event_log.reset_index("vehicle_id", drop=True)
+        .loc[:, "event_type"]
+        .map(dict(submission=2, pickup=-1, dropoff=-1))
+        .fillna(0)
+        .sort_index()
+        .groupby("timestamp")
+        .sum()
+        .cumsum()
+    )
+
+    median_system_stoplist_length = system_stoplist_length_ensemble.median()
+    mean_system_stoplist_length = system_stoplist_length_ensemble.mean()
 
     avg_detour = requests["inferred", "relative_travel_time"].mean()
 
@@ -183,6 +193,9 @@ def get_system_quantities(
         avg_waiting_time=avg_waiting_time,
         rejection_ratio=rejection_ratio,
         median_stoplist_length=median_stoplist_length,
+        mean_stoplist_length=mean_stoplist_length,
+        median_system_stoplist_length=median_system_stoplist_length,
+        mean_system_stoplist_length=mean_system_stoplist_length,
         avg_detour=avg_detour,
     )
 
