### 2024-02-09 61bd951
refactored perform_single_analysis and fixed documentation
