+++ src/ridepy/extras/simulation_set.py
@@ -65,15 +65,56 @@ def get_params(directory: Path, sim_id: str, param_path_suffix: str = "_params.j
 def perform_single_analysis(
     sim_id: str,
     data_dir: Path,
+    tasks_if_not_existent: Optional[Iterable[str]] = None,
+    tasks_if_existent: Optional[Iterable[str]] = None,
     param_path_suffix: str = "_params.json",
     event_path_suffix: str = ".jsonl",
-    update_existing: bool = False,
-    compute_system_quantities: bool = True,
-    compute_vehicle_quantities: bool = True,
     stops_path_suffix: str = "_stops.pq",
     requests_path_suffix: str = "_requests.pq",
     vehicle_quantities_path_suffix: str = "_vehicle_quantities.pq",
-) -> tuple[str, dict]:
+) -> tuple[str, dict[str, Any]]:
+    """
+    Compute stops, requests, vehicle quantities, and system quantities from
+    simulation events.
+
+    Parameters
+    ----------
+    sim_id
+        Simulation ID
+    data_dir
+        Directory from which to read the events and to which to write the resulting
+        parquet output files
+    tasks_if_not_existent
+        Collection of tasks to perform if the respective output files do not exist.
+        Valid entries are 'stops', 'requests', 'vehicle_quantities', and
+        'system_quantities'. If None, everything is computed if not existent.
+    tasks_if_existent
+        Collection of tasks to perform even if the respective output files do exist,
+        i.e. tasks to recompute. Valid entries are 'stops', 'requests',
+        'vehicle_quantities', and 'system_quantities'. If None, nothing is recomputed.
+    param_path_suffix
+        Appending this suffix to the simulation ID yields the path to the JSON parameter
+        file.
+    event_path_suffix
+        Appending this suffix to the simulation ID yields the path to the JSONL
+        event file.
+    stops_path_suffix
+        Appending this suffix to the simulation ID yields the path to the parquet
+        stops file.
+    requests_path_suffix
+        Appending this suffix to the simulation ID yields the path to the parquet
+        requests file.
+    vehicle_quantities_path_suffix
+        Appending this suffix to the simulation ID yields the path to the \
+        parquet vehicle quantities file.
+
+    Returns
+    -------
+    sim_id
+        Simulation ID
+    system_quantities
+        Dictionary containing the computed system quantities as entries
+    """
     stops_path = make_file_path(sim_id, data_dir, stops_path_suffix)
     requests_path = make_file_path(sim_id, data_dir, requests_path_suffix)
     vehicle_quantities_path = make_file_path(
@@ -82,16 +123,27 @@ def perform_single_analysis(
 
     tasks = set()
 
-    if update_existing:
-        tasks |= {"stops", "requests", "vehicle_quantities", "system_quantities"}
-    if not stops_path.exists():
+    tasks_if_not_existent = tasks_if_not_existent or {
+        "stops",
+        "requests",
+        "vehicle_quantities",
+        "system_quantities",
+    }
+    tasks_if_existent = tasks_if_existent or set()
+
+    if "stops" in tasks_if_not_existent and not stops_path.exists():
         tasks.add("stops")
-    if not requests_path.exists():
+    if "requests" in tasks_if_not_existent and not requests_path.exists():
         tasks.add("requests")
-    if compute_system_quantities:
-        tasks.add("system_quantities")
-    if compute_vehicle_quantities and not vehicle_quantities_path.exists():
+    if (
+        "vehicle_quantities" in tasks_if_not_existent
+        and not vehicle_quantities_path.exists()
+    ):
         tasks.add("vehicle_quantities")
+    if "system_quantities" in tasks_if_not_existent:
[…]
