+++ doc/extras.rst
@@ -1,12 +1,7 @@
 Module ``extras``
 =================
 
-This package contains additional functionality that is not a core part of the simulator
-but rather makes it more convenient to set up and use it. Currently this involves code
-for performing multiple simulations while varying parameters ("parameter scan"), reading
-and writing parameter sets and simulation output to and from disk as JSON/`JSON Lines`_.
-In addition, convenience methods for easily creating graph transport spaces are
-included.
+This package contains additional functionality that is not a core part of the simulator but rather makes it more convenient to set up and use it. Currently this involves code for performing multiple simulations while varying parameters ("parameter scan"), reading and writing parameter sets and simulation output to and from disk as JSON/`JSON Lines`_. In addition, convenience methods for easily creating graph transport spaces are included.
 
 
 Simulations and Parameter Scans
@@ -69,31 +64,31 @@ not specified through arguments. The order of precedence is, last taking highest
 ``default_base_params``, ``base_params``, ``zip_params``, ``product_params``.
 
 
-Executing Simulations
+Executing simulations
 ~~~~~~~~~~~~~~~~~~~~~
 
-Simulations are executed when `SimulationSet.run()` is called. Independent simulations
-are performed through executing `.perform_single_simulation()` for each parameter set
-using multiprocessing. The events that are generated by the simulation are written to
-disk in the `JSON Lines`_ format. The simulation parameters are also written to disk, in
-separate JSON files.  This includes all data necessary to perform the respective
-simulation. For more detail, see :ref:`JSON IO`.  For each simulation run, a unique
-identfier is generated and the data is stored to ``<uuid>.jsonl`` for the events and
-``<uuid>_params.json`` for the simulation parameters. The identifier hashes the
-parameter set, thereby allowing to continue an interrupted simulation set run later.
-The IDs generated can be retrieved using `SimulationSet.simulation_ids`. Alternatively the
-filenames of the resulting JSONL/JSON files are also directly available through
-`SimulationSet.param_paths` and `SimulationSet.result_paths`.
+Simulations are executed when `SimulationSet.run()` is called. Independent simulations are performed through executing `.perform_single_simulation()` for each parameter set using multiprocessing. The events that are generated by the simulation are written to disk in the `JSON Lines`_ format. The simulation parameters are also written to disk, in separate JSON files.  This includes all data necessary to perform the respective simulation. For more detail, see :ref:`JSON IO`.  For each simulation run, a unique identfier is generated and the data is stored to ``<uuid>.jsonl`` for the events and ``<uuid>_params.json`` for the simulation parameters. The identifier hashes the parameter set, thereby allowing to continue an interrupted simulation set run later. The IDs generated can be retrieved using `SimulationSet.simulation_ids`. Alternatively the filenames of the resulting JSONL/JSON files are also directly available through `SimulationSet.param_paths` and `SimulationSet.event_paths`.
 
 
-.. automodule:: ridepy.extras.simulation_set
-    :members: SimulationSet
+.. autoclass:: SimulationSet
 
-.. automodule:: ridepy.extras.simulation_set
-    :members: perform_single_simulation
+    .. automethod:: run
 
-.. automodule:: ridepy.extras.simulation_set
-    :members: simulate_parameter_combinations
+    .. autoattribute:: simulation_ids
+
+    .. autoattribute:: param_paths
+
+    .. autoattribute:: event_paths
+
+.. autofunction:: perform_single_simulation
+
+
+Running analysis on the simulation results
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The simulation results can be automatically analyzed using the `ridepy.extras.analytics` module, storing the results to disk.
+
+.. automethod:: SimulationSet.run_analytics
 
 JSON IO
 -------
