+++ doc/extras.rst
@@ -1,7 +1,7 @@
 Module ``extras``
 =================
 
-This package contains additional functionality that is not a core part of the simulator but rather makes it more convenient to set up and use it. Currently this involves code for performing multiple simulations while varying parameters ("parameter scan"), reading and writing parameter sets and simulation output to and from disk as JSON/`JSON Lines`_. In addition, convenience methods for easily creating graph transport spaces are included.
+This package contains additional functionality that is not a core part of the simulator but rather makes it more convenient to set up and use it. Currently this involves code for performing multiple simulations while varying parameters ("parameter scan"), reading and writing parameter sets and simulation output to and from disk as `JSON`_/`JSON Lines`_. In addition, convenience methods for easily creating graph transport spaces are included.
 
 
 Simulations and Parameter Scans
@@ -29,17 +29,7 @@ parameters, i.e. performing a parameter scan.  The typical workflow is as follow
 Parameter Configuration
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-`SimulationSet` takes three main arguments: ``base_params``, ``zip_params`` and
-``product_params``. Base parameters are parameters which are kept constant across all
-simulations defined by the simulation set. Here the values of the inner dict are the
-actual parameters. For zip and product parameters, lists of values are supplied as the
-inner dictionary's values. Zip parameters are varied simultaneously across the
-simulations, i.e. the first simulation will use the first parameter value for all of the
-parameters in ``zip_params``, the second simulation will use the second parameter
-values, and so on. For zip paramaters it is important that all lists of parameter values
-are of equal length. The lists in product parameters on the other hand will be
-multiplied as a Cartesian product. Here the lengths do not have to match, all possible
-combinations will be simulated.
+`SimulationSet` takes three main arguments: ``base_params``, ``zip_params`` and ``product_params``. Base parameters are parameters which are kept constant across all simulations defined by the simulation set. Here the values of the inner dict are the actual parameters. For zip and product parameters, lists of values are supplied as the inner dictionary's values. Zip parameters are varied simultaneously across the simulations, i.e. the first simulation will use the first parameter value for all of the parameters in ``zip_params``, the second simulation will use the second parameter values, and so on. For zip paramaters it is important that all lists of parameter values are of equal length. The lists in product parameters on the other hand will be multiplied as a Cartesian product. Here the lengths do not have to match, all possible combinations will be simulated.
 
 The current list of supported parameters is:
 
@@ -67,7 +57,7 @@ not specified through arguments. The order of precedence is, last taking highest
 Executing simulations
 ~~~~~~~~~~~~~~~~~~~~~
 
-Simulations are executed when `SimulationSet.run()` is called. Independent simulations are performed through executing `.perform_single_simulation()` for each parameter set using multiprocessing. The events that are generated by the simulation are written to disk in the `JSON Lines`_ format. The simulation parameters are also written to disk, in separate JSON files.  This includes all data necessary to perform the respective simulation. For more detail, see :ref:`JSON IO`.  For each simulation run, a unique identfier is generated and the data is stored to ``<uuid>.jsonl`` for the events and ``<uuid>_params.json`` for the simulation parameters. The identifier hashes the parameter set, thereby allowing to continue an interrupted simulation set run later. The IDs generated can be retrieved using `SimulationSet.simulation_ids`. Alternatively the filenames of the resulting JSONL/JSON files are also directly available through `SimulationSet.param_paths` and `SimulationSet.event_paths`.
[…]
