### 2024-02-09 835df6f
241 improve analytics documentation (#242)
* moved lru cache module, added analytics modules to sphinx

* added system stoplist length computation

* refactored perform_single_analysis and fixed documentation

* added blacken-docs for doc code formatting

* json links

* intro page now mentions repository
