+++ test/test_analytics.py
@@ -902,6 +902,9 @@ def test_get_stops_and_requests_and_get_quantities():
             "efficiency_time": (0.3 + 0.1 + 1) / (0.1 + 0.1 + 0.1 + 1 + 1),
             "avg_waiting_time": (0 + 0.1 + 1) / 3,
             "median_stoplist_length": np.median([3, 2, 1, 0, 0, 2, 1, 0, 0, 0]),
+            "mean_stoplist_length": np.mean([3, 2, 1, 0, 0, 2, 1, 0, 0, 0]),
+            "median_system_stoplist_length": np.median([5, 4, 3, 2, 1, 0]),
+            "mean_system_stoplist_length": np.mean([5, 4, 3, 2, 1, 0]),
             "rejection_ratio": 0.25,
             "avg_detour": 1.0,
             "avg_system_stoplist_length_service_time": 1.4666666666666668,
