+++ doc/analytics.rst
@@ -1,0 +1,8 @@
+The analytics package
+---------------------
+
+.. automodule:: ridepy.util.analytics
+    :members: get_stops_and_requests, get_stops_and_requests_from_events_dataframe, get_vehicle_quantities, get_system_quantities
+    :undoc-members:
+    :show-inheritance:
+

[…]
