### 2024-02-09 06acf6d
ridepy 2.2.1.post1
Typo in documentation resolved.
