### 2024-02-16 409560e
added some properties including direct system quantities access to simulation set
