### 2024-02-20 3ad84c5
247 add convenience function to join extra parameters to the analytics system quantities output (#248)
* added some properties including direct system quantities access to simulation set

* enhanced test

* Update index.rst
