+++ doc/index.rst
@@ -3,7 +3,7 @@ Welcome to RidePy's documentation!
 
 RidePy is a scientific Python library for simulating modern on-demand transit systems such as ridepooling.
 
-To better understand what RidePy is all about probably should start by reading the :doc:`overview section <overview>`. For instructions on installation, contributing and support, head over to the :doc:`*How to...* section <howto>`. And to dive right in and get started with using RidePy for a simple simulation, try the :doc:`tutorial <notebooks>`.
+To better understand what RidePy is all about probably should start by reading the :doc:`overview section <overview>`. For instructions on installation, contributing and support, head over to the :doc:`**How to...** section <howto>`. And to dive right in and get started with using RidePy for a simple simulation, try the :doc:`tutorial <notebooks>`.
 
 In addition to these general guides, we offer a commented :doc:`API reference <reference>`, in which we explain the building blocks and concepts that RidePy is based on in more detail, and how you can leverage them for your specific application. This section is far more technical than the hands-on tutorial, but allows for a deeper understanding.
 
