### 2024-02-20 360f989
Added CLI utility to publish ridepy releases (#250)
* implemented most of it

* should work now

* more magic

* should be done

* doc
