### 2024-03-09 #6241 @abhishektiwari
@fxjung Great work. Overall I enjoyed reviewing this software. Please see my initial comments (all of them non-blocking),

1. Please consider improving the [Github repo](https://github.com/PhysicsOfMobility/ridepy) `Readme` by adding installation instructions and contribution guidelines. I know these are already covered by [documentation portal](https://ridepy.org/) in great details but given Github repo is landing page for this software paper I highly recommend to update `Readme`.
2. Please consider  improving  the `A statement of need` and `Summary` covered by [Github repo](https://github.com/PhysicsOfMobility/ridepy)  `Readme` and [documentation portal](https://ridepy.org/) landing page, and align them with paper. I feel they all read differently and only after reading the software paper or the [conceptual overview](https://ridepy.org/overview.html) one can understand the problem statement and target audience.
3. As mentioned by paper ridepy's modular design makes customization easy, while the included modules allow for a quick start, documentation will benefit with examples of how to extend the ridepy modules with a real-world use case. API/code documentation for supported dispatching algorithms partially covers this but additional details will help users.
4. Installation instructions worked but can be little bit clear by calling out pre-installation of C++ build environment and the Boost libraries. For instance, on Debian run `sudo apt update && sudo apt -y install build-essential libboost-all-dev` before `pip install ridepy`.

I commend authors effort to document a [domain glossary](https://ridepy.org/glossary.html) to help anyone not familiar with the space to rationalise it quickly. There is already a comprehensive backlog of improvements documented in issue tracker so it seems there is additional unit tests planned to improve the test coverage.

@diehlpk Will wait for response from @fxjung, but review is mostly completed from my end. Please let me know if anything else required.
