+++ README.rst
@@ -15,9 +15,7 @@ Quickstart
 If you're in a hurry, here's the gist. For detailed instructions, see the `installation guide <https://ridepy.org/setup.html>`__ in the documentation.
 
 - RidePy currently works best with Python 3.9
-- If you are on platform that we don't offer wheels for, you will need to first set up a C++ build environment and the `Boost C++ libraries <https://www.boost.org/>`_.
-
-  On Debian-like systems, this can be easily done with ``sudo apt-get install build-essential libboost-all-dev``.
+- If you are on platform that we don't offer wheels for, you will need to first set up a C++ build environment and the `Boost C++ libraries <https://www.boost.org/>`_. On Debian-like systems, this can be easily done with ``sudo apt-get install build-essential libboost-all-dev``.
 - Install RidePy with ``pip install ridepy``.
 
 ..
