+++ README.rst
@@ -12,14 +12,16 @@ Head over to `ridepy.org <ridepy doc_>`_ to get started.
 Quickstart
 ----------
 
-If you're in a hurry, here's the gist. For detailed instructions, see the `installation guide <https://ridepy.org/setup.html>`__ in the documentation.
+For detailed instructions, see the `installation guide <https://ridepy.org/setup.html>`__ in the documentation.
+
+If you're in a hurry, here's the gist.
 
 - RidePy currently works best with Python 3.9
 - If you are on platform that we don't offer wheels for, you will need to first set up a C++ build environment and the `Boost C++ libraries <https://www.boost.org/>`_. On Debian-like systems, this can be easily done with
 
 .. code-block:: bash
 
-    sudo apt-get install build-essential libboost-all-dev``
+    sudo apt-get install build-essential libboost-all-dev
 
 - Install RidePy with
 
