+++ README.rst
@@ -17,7 +17,7 @@ For detailed instructions, see the `installation guide <https://ridepy.org/setup
 If you're in a hurry, here's the gist:
 
 - RidePy currently works best with Python 3.9
-- If you are on platform other than x86-64 Linux, i.e. one that we don't offer wheels for, you will need to first set up a C++ build environment and the `Boost C++ libraries <https://www.boost.org/>`_. On Debian-like systems, this is easily accomplished with
+- If you are on platform other than x86-64 Linux, i.e., one that we don't offer wheels for, you will need to first set up a C++ build environment and the `Boost C++ libraries <https://www.boost.org/>`_. On Debian-like systems, this is easily accomplished with
 
 .. code-block:: bash
 
