+++ README.rst
@@ -21,13 +21,45 @@ If you're in a hurry, here's the gist:
 
 .. code-block:: bash
 
-    sudo apt-get install build-essential libboost-all-dev
+   sudo apt-get install build-essential libboost-all-dev
 
 - Install RidePy with
 
 .. code-block:: bash
 
-    pip install ridepy
+   pip install ridepy
+
+Contributing
+------------
+
+We are always happy for contributions from the community. If you run into a problem, please `report an issue <https://ridepy.org/issues.html>`_ or `ask for help <https://ridepy.org/support.html>`_.
+
+If you are interested in contributing to our codebase, please read our `contributing guide <https://ridepy.org/contributing.html>`_.
+
+Here is a short overview of the most important points:
+
+Resources
+~~~~~~~~~
+
+- Documentation: `ridepy.org <ridepy doc_>`_
+- Source code: `GitHub <https://github.com/PhysicsOfMobility/ridepy>`_
+- Issue tracker: `GitHub <https://github.com/PhysicsOfMobility/ridepy/issues>`_
+
+Code style
+~~~~~~~~~~
+
+- Python: `black <https://github.com/psf/black>`_
+- C++: `LLVM <https://llvm.org/docs/CodingStandards.html>`_
+
+Development
+~~~~~~~~~~~
+
+- Version control: `Git <https://git-scm.com/>`_
+- Testing: `pytest <https://docs.pytest.org/en/stable/>`_
+- Continuous integration: `GitHub Actions <https://github.com/PhysicsOfMobility/ridepy/actions>`_
+- Versioning: `Semantic Versioning <https://semver.org/>`_
+
+.. http://mozillascience.github.io/working-open-workshop/contributing/
 
 ..
     ---------
@@ -35,18 +67,18 @@ If you're in a hurry, here's the gist:
     ---------
 
 .. |Code style: black| image:: https://img.shields.io/badge/code%20style-black-000000.svg
-    :target: https://github.com/psf/black
+   :target: https://github.com/psf/black
 
 .. |Docs| image:: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-doc.yml/badge.svg
-    :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-doc.yml
+   :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-doc.yml
 
 .. |Tests| image:: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/python-testing.yml/badge.svg
-    :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/python-testing.yml
+   :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/python-testing.yml
 
 .. |wheel| image:: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-wheel.yml/badge.svg
-    :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-wheel.yml
+   :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-wheel.yml
 
 .. |sdist| image:: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-sdist.yml/badge.svg
-    :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-sdist.yml
+   :target: https://github.com/PhysicsOfMobility/ridepy/actions/workflows/build-sdist.yml
 
 .. _ridepy doc: https://ridepy.org
