### 2024-03-11 16ed9ae
made installation prerequisites more visible
