+++ doc/howto.rst
@@ -4,8 +4,8 @@ How to...
 .. toctree::
    :maxdepth: 1
 
-    Install RidePy? <setup>
-    Contribute? <contributing>
-    Report an issue? <issues>
-    Obtain support? <support>
+    Install RidePy <setup>
+    Contribute <contributing>
+    Report an issue <issues>
+    Obtain support <support>
 
