### 2024-03-11 1ae879a
made installation prerequisites more visible
