+++ src/ridepy/util/dispatchers/helper_functions.py
@@ -1,202 +1,0 @@
-from typing import Optional
-from copy import deepcopy
-
-from ridepy.data_structures import (
-    Stoplist,
-    TransportationRequest,
-    TransportSpace,
-    Stop,
-    StopAction,
-)
-
-
-def insert_request_to_stoplist_drive_first(
-    stoplist: Stoplist,
-    request: TransportationRequest,
-    pickup_idx: int,
-    dropoff_idx: int,
-    space: TransportSpace,
-    n_passengers: int = 1,  # future proofing when we have requests with multiple passengers
-) -> Stoplist:
-    """
-    Inserts a request into  a stoplist. The pickup (dropoff) is inserted after pickup_idx (dropoff_idx).
-    The estimated arrival times at all the stops are updated assuming a drive-first strategy.
-    """
-    # We don't want to modify stoplist in place. Make a copy.
-    new_stoplist = deepcopy(stoplist)
-
-    # Handle the pickup
-    stop_before_pickup = new_stoplist[pickup_idx]
-    cpat_at_pu = stop_before_pickup.estimated_departure_time + space.t(
-        stop_before_pickup.location, request.origin
-    )
-    pickup_stop = Stop(
-        location=request.origin,
-        action=StopAction.pickup,
-        estimated_arrival_time=cpat_at_pu,
-        time_window_min=request.pickup_timewindow_min,
-        time_window_max=request.pickup_timewindow_max,
-        request=request,
-        occupancy_after_servicing=stop_before_pickup.occupancy_after_servicing
-        + n_passengers,
-    )
-    # increase the occupancies of all the stops between pickup and dropoff
-    # remember, the indices are as follows:
-    # 0,1,...,pickup_idx,(pickup_not_yet_inserted),...,dropoff_idx,(dropoff_not_yet_inserted), ...
-    for s in new_stoplist[pickup_idx + 1 : dropoff_idx + 1]:
-        s.occupancy_after_servicing += n_passengers
-
-    insert_stop_to_stoplist_drive_first(new_stoplist, pickup_stop, pickup_idx, space)
-
-    # Handle the dropoff
-    dropoff_idx += 1
-    stop_before_dropoff = new_stoplist[dropoff_idx]
-    cpat_at_do = stop_before_dropoff.estimated_departure_time + space.t(
-        stop_before_dropoff.location, request.destination
-    )
-    dropoff_stop = Stop(
-        location=request.destination,
-        action=StopAction.dropoff,
-        estimated_arrival_time=cpat_at_do,
-        time_window_min=request.delivery_timewindow_min,
-        time_window_max=request.delivery_timewindow_max,
-        request=request,
-        occupancy_after_servicing=stop_before_dropoff.occupancy_after_servicing
-        - n_passengers,
-    )
-    insert_stop_to_stoplist_drive_first(new_stoplist, dropoff_stop, dropoff_idx, space)
-
-    return new_stoplist
-
-
-def insert_stop_to_stoplist_drive_first(
-    stoplist: Stoplist,
-    stop: Stop,
-    idx: int,
-    space: TransportSpace,
-) -> None:
-    """
-    Note: Modifies stoplist in-place. The passed stop has estimated_arrival_time set to None
-    Args:
-        stoplist:
-        stop:
-        idx:
-        space:
-
-    Returns:
-    """
-    stop_before_insertion = stoplist[idx]
-    stop.estimated_arrival_time = cpat_of_inserted_stop(
-        stop_before=stop_before_insertion,
-        time_from_stop_before=space.t(stop_before_insertion.location, stop.location),
-    )
-
-    if idx < len(stoplist) - 1:
-        # update CPATs of later stops
-        delta_CPAT_next_stop = (
-            stop.estimated_departure_time
-            + space.t(stop.location, stoplist[idx + 1].location)
[…]
