+++ src/ridepy/util/dispatchers/ridepooling.py
@@ -6,7 +6,7 @@ from ridepy.data_structures import (
     TransportationRequest,
     Stoplist,
     TransportSpace,
-    SingleVehicleSolution,
+    DispatcherSolution,
     Stop,
     StopAction,
 )
@@ -215,7 +215,7 @@ def BruteForceTotalTravelTimeMinimizingDispatcher(
     stoplist: Stoplist,
     space: TransportSpace,
     seat_capacity: int,
-) -> SingleVehicleSolution:
+) -> DispatcherSolution:
     """
     Dispatcher that maps a vehicle's stoplist and a request to a new stoplist
     by minimizing the total driving time.
