+++ doc/dispatcher-guide.rst
@@ -1,4 +1,271 @@
 Writing a dispatcher
 ====================
 
-gt
+General overview
+----------------
+
+Dispatchers are an essential part of RidePy: They determine the routes that the vehicles are going to take.
+
+In essence, a dispatcher is a mapping of a new request and an existing stoplist associated with a single vehicle to an updated stoplist and a cost of inserting the request:
+
+.. math::
+
+   (\text{new request}, \text{stoplist})\mapsto(\text{cost}, \text{new stoplist})
+
+The dispatcher is responsible for the following steps:
+
+* Checking whether the pickup and dropoff stops for the new request may be inserted into the stoplist without violating any constraints.
+* These constraints may entail:
+  * Time window constraints of the request's pick-up and drop-off stops
+  * Time window constraints of all other stops already in the stoplist
+  * Vehicle seat capacity constraints
+  * The trivial ridepooling constraint, i.e., the pick-up location needs to be visited before the drop-off location
+* Multiple feasible solutions may exist. In this case, the dispatcher is expected to determine the solution ("insertion") incurring the minimum cost by some chose metric (the value of which is returned).
+* It might happen that no solution can be found. In this case, the dispatcher must return a float :math:`\text{cost} = \infty` to let the caller know that the request can't be serviced by the vehicle in question.
+
+It must be stressed that the dispatcher is solely responsible for checking the aforementioned constraints and for ensuring that the stoplists are kept in a valid state. This also means that the dispatcher must update the estimated arrival times of all stops except for the CPE (current position element)  which stores the current or last known location of the vehicle and is updated by `.VehicleState`. The dispatcher must not delete it, though.
+
+The dispatcher is called on every vehicle's stoplist to determine the vehicle and route that incurs the least cost. If the minimum cost is infinite, i.e., the dispatcher has failed to return a finite cost of insertion for any of the vehicles, the rejection is *rejected*.
+
+Before starting to write your own dispatcher, it might prove helpful to have a look at the source code of the trivial Python `.TaxicabDispatcherDriveFirst` and possibly that of the  Python `.BruteForceTotalTravelTimeMinimizingDispatcher` to get some hints.
+
+Signature
+---------
+
+In the most simple case, a RidePy dispatcher is a Python function with the following signature:
+
+.. code-block:: python
+
+   @dispatcherclass
+   def MyDispatcher(
+       request: TransportationRequest,
+       stoplist: Stoplist,
+       space: TransportSpace,
+       seat_capacity: int,
+   ) -> DispatcherSolution: ...
+
+Let's disect this:
+
+- ``MyDispatcher`` is a pure function taking a `.TransportationRequest`, a `.Stoplist`, a `.TransportSpace` and an integer seat capacity
+- A `.TransportationRequest` is defined like this:
+
+.. code-block:: python
+
+   ID = Union[str, int]
+   """Generic ID, could be vehicle ID, request ID, ..."""
+
+
+   @dataclass
+   class Request:
+       """
+       A request for the system to perform a task
+       """
+
+       request_id: ID
+       creation_timestamp: float
+
+
+   @dataclass
+   class TransportationRequest(Request):
+       """
+       A request for the system to perform a transportation task,
+       through creating a route through the system given spatio-temporal constraints.
+       """
+
+       origin: Any
+       destination: Any
+       pickup_timewindow_min: float = 0
+       pickup_timewindow_max: float = inf
+       delivery_timewindow_min: float = 0
+       delivery_timewindow_max: float = inf
+
+- A `.Stoplist` is a list of `.Stop` objects:
+
+.. code-block:: python
+
+   Stoplist = List[Stop]
[…]
