+++ doc/dispatcher-guide.rst
@@ -1,8 +1,8 @@
 Writing a dispatcher
 ====================
 
-General overview
-----------------
+General information
+-------------------
 
 Dispatchers are an essential part of RidePy: They determine the routes that the vehicles are going to take.
 
@@ -16,10 +16,12 @@ The dispatcher is responsible for the following steps:
 
 * Checking whether the pickup and dropoff stops for the new request may be inserted into the stoplist without violating any constraints.
 * These constraints may entail:
+
   * Time window constraints of the request's pick-up and drop-off stops
   * Time window constraints of all other stops already in the stoplist
   * Vehicle seat capacity constraints
   * The trivial ridepooling constraint, i.e., the pick-up location needs to be visited before the drop-off location
+
 * Multiple feasible solutions may exist. In this case, the dispatcher is expected to determine the solution ("insertion") incurring the minimum cost by some chose metric (the value of which is returned).
 * It might happen that no solution can be found. In this case, the dispatcher must return a float :math:`\text{cost} = \infty` to let the caller know that the request can't be serviced by the vehicle in question.
 
@@ -27,10 +29,25 @@ It must be stressed that the dispatcher is solely responsible for checking the a
 
 The dispatcher is called on every vehicle's stoplist to determine the vehicle and route that incurs the least cost. If the minimum cost is infinite, i.e., the dispatcher has failed to return a finite cost of insertion for any of the vehicles, the rejection is *rejected*.
 
-Before starting to write your own dispatcher, it might prove helpful to have a look at the source code of the trivial Python `.TaxicabDispatcherDriveFirst` and possibly that of the  Python `.BruteForceTotalTravelTimeMinimizingDispatcher` to get some hints.
+Before starting to write your own dispatcher, it might prove helpful to have a look at the source code of the fairly trivial Python `.TaxicabDispatcherDriveFirst` dispatcher and possibly that of the Python `.BruteForceTotalTravelTimeMinimizingDispatcher` dispatcher to get some hints.
 
-Signature
----------
+Python vs Cython/C++ dispatchers
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+RidePy dispatchers may be written either in Python or in Cython/C++. The latter is recommended for performance-critical dispatchers. This guide will focus on Python dispatchers for now.
+
+If you nonetheless want to have a look at the Cython/C++ dispatchers, you can find the currently implemented ones in the `ridepy.util.dispatchers_cython` module. There, the actual dispatching logic is implemented in ``cdispatchers.h``, which is exposed to Cython/Python by ``cdispatchers.pxd`` and ``cdispatchers.pyx``. In ``dispatchers.pyx`` and ``dispatchers.pxd``, an additional layer allows to set the `.LocType` appropriate for the `.TransportSpace` that is used. If this section confused you, please ignore it for now.
+
+.. _dispatcherclass:
+
+Function vs class dispatchers
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Strictly speaking, dispatchers in RidePy are not pure functions, but classes. This means, dispatchers can have state, such as the `.TransportSpace`'s `.LocType` for Cython/C++ dispatchers. This does not need to bother you at the moment, though, as there is a convenient wrapper for Python dispatchers that allows you to write them as pure functions. Just wrap your pure function dispatcher with the `.dispatcherclass` decorator from `.ridepy.util.dispatchers.dispatcher_class` like below, and you're good to go.
+
+
+Signature of a Python dispatcher
+--------------------------------
 
 In the most simple case, a RidePy dispatcher is a Python function with the following signature:
 
[…]
