### 2024-03-12 539b02d
bumped black, fixed broken error logic
