+++ pyproject.toml
@@ -53,7 +53,7 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
-    "black==24.1.1",
+    "black==24.2.0",
     "commitizen",
     "pdbpp",
     "pre-commit",
