### 2024-03-12 6a8c0a2
Implement JOSS reviewer comments regarding documentation (#253)
* readme changed

* readme changed

* Update README.rst

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* readme changed

* made installation prerequisites more visible

* made installation prerequisites more visible

* slightly restructured py dispatcher structure + bugfix

* started dispatcher writing guide

* guide almost done

* glossary

* glossary

* guide almost done

* bumped black, fixed bro
[…]
