### 2024-03-22 #6241 @fxjung
@diehlpk Sorry I didn't mention this in my reply above---yes, I have tried to address them: I have added quick-start and contribution sections to the README, a guide in the docs on how to write a custom dispatcher, and I have improved the installation instructions.

Regarding the second comment I was not really sure what to do as I pointed out in my earlier reply.
