### 2024-03-27 #6241 @abhishektiwari
@fxjung Thanks for revisions and update.

Re: Comment-2 Please consider comment it resolved from my side. I see you already have revised the statement of need/summary on Readme and documentation portal.
Re: comment-3 Details documentation on how to write a [custom dispatcher](https://ridepy.org/dispatcher-guide.html) looks great. Thanks for putting extra effort there. Users will love it.
Re: Comment-1 and Comment-4 are also addressed.

@diehlpk No open issues or comments from my side.
