### 2024-03-27 #6241 @diehlpk
> ### Paper
>
> The opening summary section/paragraph could probably benefit from some description of analyses enabled by the software, mainly for non-specialist readers. This is my only hangup on the checklist, though arguably the "Philosopy and usage" section covers it.
>
> Also in the summary section, there's a distinction made between "modeling the mobility service itself rather than its customers or the environment." I don't think a full justification fits here, but it could be helpful to point out what benefits this design decision has, whether it improves modeling flexibility, "naturalness" of model specification, performance, etc. I could imagine the sentence continuing like "...rather than its customers or environment, enabling ..."
>
> ~The citation of the Julia language seems superfluous to me, since it's only mentioned in reference to another package.~ The package is also referred to as `RidePooling.jl`, but I'm not seeing any package by that name. I think the description "...developed in support of a recent scientific contribution" is vague enough to not really be useful. If the idea is to point out that it's academic (what I'm guessing from "idealized approach"), it might be good to specify some feature(s) RidePy offers, or something to that effect.

@fxjung can you please have a look at this comment?
