### 2024-04-08 4f9b276
added request generator interface
