+++ doc/request-generator-guide.rst
@@ -1,0 +1,4 @@
+Writing a request generator
+===========================
+
+Request generators are iterators
