### 2024-04-08 2e96956
Added request generator interface (#255)
