+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 -   repo: https://github.com/python/black
-    rev: 24.1.1
+    rev: 24.3.0
     hooks:
     - id: black
       language_version: python3.9
@@ -9,4 +9,4 @@ repos:
     hooks:
     - id: blacken-docs
       additional_dependencies:
-        - black==24.2.0
+        - black==24.3.0
