### 2024-04-09 17cc80a
Upgraded napoleon to sphinx-bundled version
