+++ pyproject.toml
@@ -77,7 +77,6 @@ doc = [
     "recommonmark",
     "sphinx",
     "sphinx-rtd-theme",
-    "sphinxcontrib-napoleon",
     "sphinx-toggleprompt",
     "sphinx-autodoc-typehints",
     "myst-nb",
