### 2024-04-09 a82cdf8
Added DistanceDistribution interface (#256)
* Upgraded napoleon to sphinx-bundled version

* Added DistanceDistribution abc
