### 2024-04-10 4854fd7
Having no accepted requests no longer crashes analytics
