### 2024-04-10 270c65e
Modified groupby calls according to pandas FutureWarning
