### 2024-04-10 81168b5
Having no accepted requests no longer crashes analytics, part II
