### 2024-04-10 47762de
Having no accepted requests no longer crashes analytics (#257)
* Having no accepted requests no longer crashes analytics

* Modified groupby calls according to pandas FutureWarning

* Having no accepted requests no longer crashes analytics, part II
