### 2024-04-16 6f8a1ec
Small enhancement to the manuscript summary (#258)
* Small enhancement to the summary

* Update paper.md
