+++ joss/paper.bib
@@ -243,36 +243,6 @@
   file = {/home/jung/Zotero/storage/A2GDNDIG/Molkenthin et al. - 2020 - Scaling Laws of Collective Ride-Sharing Dynamics.pdf;/home/jung/Zotero/storage/BMY6D6F8/PhysRevLett.125.html}
 }
 
-@misc{muehle2022,
-  title = {{{RidePoolingSimulations}}},
-  author = {M{\"u}hle, Steffen},
-  year = {2022},
-  month = jun,
-  url = {https://github.com/SteffenMuehle/RidePoolingSimulations},
-  urldate = {2023-04-25},
-  abstract = {Simulation code, animations and results from the paper "A framework for modeling ride pooling efficiency and minimum fleet size"},
-  copyright = {MIT}
-}
-
-@article{muehle2023,
-  ids = {muhle2023},
-  title = {An Analytical Framework for Modeling Ride Pooling Efficiency and Minimum Fleet Size},
-  author = {M{\"u}hle, Steffen},
-  year = {2023},
-  month = jun,
-  journal = {Multimodal Transportation},
-  volume = {2},
-  number = {2},
-  pages = {100080},
-  issn = {2772-5863},
-  doi = {10.1016/j.multra.2023.100080},
-  urldate = {2023-03-02},
-  abstract = {Ride pooling (RP) is a transport mode using on-demand buses to combine the trips of multiple users into one vehicle. Its required fleet size and carbon emissions are quantified by the system's efficiency. Due to the complex interplay between street network, buses, users and dispatch algorithm, efficiency case studies are available but bottom-up predictions are not. Here we close this gap using probabilistic and geometric arguments in an analytical model framework. Its modular design allows for adaptation to specific usage scenarios and provides an over-arching view of them. In a showcase on Euclidean spaces, our model quantifies how RP outperforms private cars as user demand increases. Its predicted power-law scaling is verified using a custom simulation framework, which further reveals improved scaling on real street networks and graphs with hierarchical structures. Henceforth, our work may help to identify street networks well-suited for RP, and predict key performance indicators analytically.},
-  langid = {english},
-  keywords = {Demand-responsive transport,Mobility as a service,Ride pooling,Ride sharing,Street network,Sustainability},
-  file = {/home/jung/Zotero/storage/4MHIX7EG/Mühle - 2023 - An analytical framework for modeling ride pooling .pdf;/home/jung/Zotero/storage/FITLHAI4/S2772586323000126.html}
-}
-
 @misc{pytest,
   title = {Pytest},
   author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
