+++ doc/howto.rst
@@ -1,11 +1,0 @@
-How to...
-=========
-
-.. toctree::
-   :maxdepth: 1
-
-   Install RidePy <setup>
-   Contribute <contributing>
-   Report an issue <issues>
-   Obtain support <support>
-   Write a dispatcher <dispatcher-guide>
