+++ doc/howto/index.rst
@@ -1,0 +1,11 @@
+How to...
+=========
+
+.. toctree::
+   :maxdepth: 1
+
+   Install RidePy <install-ridepy>
+   Contribute <contribute>
+   Report an issue <report-an-issue>
+   Obtain support <obtain-support>
+   Write a dispatcher <write-a-dispatcher>
