### 2024-04-18 10d5dfc
added doc on how to release
