### 2024-04-18 88d1382
updated doc on how to release
