### 2024-04-18 cdc4e2d
move doc build to dirhtml
