+++ .github/workflows/build-doc.yml
@@ -33,6 +33,6 @@ jobs:
           pip install -U pip 
           pip install .[doc]
       - name: Build doc
-        run: sphinx-build -b html ./doc/ ./_site
+        run: sphinx-build -b dirhtml ./doc/ ./_site
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
