### 2024-04-25 #6241 @diehlpk
@fxjung

1. Dresden **(cfaed)**, TUD Dresden University of Technology, 01062 Dresden -> Should be CFAED in capital letters?
2. Chair of Network Dynamics -> This reads strange to me. Are you both the department chair? Maybe the institute's name is sufficient?
