### 2024-04-26 #6241 @fxjung
@diehlpk Thanks for checking in regarding this. The affiliation is correct, however. "cfaed" is officially lower-case, and "Chair of Network Dynamics" refers to the "Lehrstuhl", which is part of our standard affiliation.

Do I need to act on any of the items in the post-review checklist?
