+++ joss/paper.md
@@ -26,10 +26,10 @@ RidePy provides fast computer simulations of on-demand mobility modes such as ri
 
 # Statement of need
 
-An accelerating climate change and congested cities both call for an urgent change in the way we move [@winkler2023]. To reduce carbon dioxide emissions as well as the number of vehicles on the road, digitally managed on-demand mobility services such as ridehailing and ridepooling are explored in research [@santi2014; @engelhardt2019] and on the road. Unfortunately, physically experimenting with such services for research purposes is extremely cost- and time-intensive. However, the operational properties of such systems are largely predefined in terms of the scheduling backend that manages them. This makes it possible to replace physical experiments with computer simulations, substituting virtual vehicles for actual ones and modeling the incoming mobility demand by sampling either historic requests or synthetic distributions. Another advantage of simulations is that the degree to which they represent reality may be freely adjusted. This makes it possible to both answer concrete operational questions [@henao2019a; @ruch2020; @wilkes2021a; @zwick2021a; @zwick2022c; @lotze2022b; @deruijter2023] and investigate idealized system behavior, gaining deeper insights into the general properties of on-demand mobility systems [@tachet2017a; @herminghaus2019a; @molkenthin2020; @manik2020a; @zech2022].
+The accelerating climate change and congested cities both call for an urgent change in the way we move [@winkler2023]. To reduce carbon dioxide emissions as well as the number of vehicles on the road, digitally managed on-demand mobility services such as ridehailing and ridepooling are explored in research [@santi2014; @engelhardt2019] and on the road. Unfortunately, physically experimenting with such services for research purposes is extremely cost- and time-intensive. However, the operational properties of such systems are largely predefined in terms of the scheduling backend that manages them. This makes it possible to replace physical experiments with computer simulations, substituting virtual vehicles for actual ones and modeling the incoming mobility demand by sampling either historic requests or synthetic distributions. Another advantage of simulations is that the degree to which they represent reality may be freely adjusted. This makes it possible to both answer concrete operational questions [@henao2019a; @ruch2020; @wilkes2021a; @zwick2021a; @zwick2022c; @lotze2022b; @deruijter2023] and investigate idealized system behavior, gaining deeper insights into the general properties of on-demand mobility systems [@tachet2017a; @herminghaus2019a; @molkenthin2020; @manik2020a; @zech2022].
 
 In this context, a simulation framework should appropriately allow for vastly different system sizes and degrees of realism. The system size incorporates the number of simulated vehicles as well as the extent of the space they operate on: A small system may consist of a single vehicle serving a network of just two nodes, while an example of a large system could be a fleet of several thousand vehicles operating on the street network of a large city. The degree of realism may be varied, for example, by sampling requests from either a uniform distribution or recorded mobility demand, or by operating on a continuous Euclidean plane versus a realistic city street network. 
-Another option is to adjust the constraints imposed, such as the time windows assigned to stops or the vehicles' seat capacities. 
+Another option is to adjust the constraints imposed, such as the time windows assigned to stops, or the vehicles' seat capacities. 
 
 Finally, an on-demand mobility simulation framework should be fast, easy to use and adaptable to various applications.
 
@@ -43,9 +43,9 @@ RidePy extends this landscape by providing a universal and fast ridepooling simu
 
 # Philosophy and usage
 
[…]
