### 2024-04-29 bba9717
add manuscript files
