### 2024-04-29 5551e87
Add JOSS manuscript to doc (#259)
