+++ doc/joss/paper.bib
@@ -1,0 +1,474 @@
+@book{behrisch2014,
+  title = {Simulation of {{Urban Mobility}}: {{First International Conference}}, {{SUMO}} 2013, {{Berlin}}, {{Germany}}, {{May}} 15-17, 2013. {{Revised Selected Papers}}},
+  shorttitle = {Simulation of {{Urban Mobility}}},
+  editor = {Behrisch, Michael and Krajzewicz, Daniel and Weber, Melanie},
+  year = {2014},
+  series = {Information {{Systems}} and {{Applications}}, Incl. {{Internet}}/{{Web}}, and {{HCI}}},
+  edition = {1st ed. 2014},
+  number = {8594},
+  publisher = {{Springer Berlin Heidelberg : Imprint: Springer}},
+  address = {{Berlin, Heidelberg}},
+  doi = {10.1007/978-3-662-45079-6},
+  abstract = {This book constitutes the thoroughly refereed proceedings of the First International Conference on Simulation of Urban Mobility, SUMO 2013, held in Berlin, Germany, in May 2013. The 12 revised full papers presented tin this book were carefully selected and reviewed from 22 submissions. The papers are organized in two topical sections: models and technical innovations and applications and surveys},
+  isbn = {978-3-662-45079-6},
+  lccn = {003.3},
+  keywords = {Application software,Computer Applications,Computer simulation,Simulation and Modeling},
+  file = {/home/jung/Zotero/storage/J4B375AA/Behrisch et al. - 2014 - Simulation of Urban Mobility First International .pdf}
+}
+
+@article{bezanson2017,
+  title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
+  shorttitle = {Julia},
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  year = {2017},
+  month = jan,
+  journal = {SIAM Rev.},
+  volume = {59},
+  number = {1},
+  pages = {65--98},
+  issn = {0036-1445, 1095-7200},
+  doi = {10.1137/141000671},
+  urldate = {2023-04-25},
+  langid = {english},
+  file = {/home/jung/Zotero/storage/SRV952D3/Bezanson et al. - 2017 - Julia A Fresh Approach to Numerical Computing.pdf}
+}
+
+@article{cython,
+  ids = {behnel2011},
+  title = {Cython: {{The Best}} of {{Both Worlds}}},
+  shorttitle = {Cython},
+  author = {Behnel, Stefan and Bradshaw, Robert and Citro, Craig and Dalcin, Lisandro and Seljebotn, Dag Sverre and Smith, Kurt},
+  year = {2011},
+  month = mar,
+  journal = {Comput. Sci. Eng.},
+  volume = {13},
+  number = {2},
+  pages = {31--39},
+  issn = {1521-9615},
+  doi = {10.1109/MCSE.2010.118},
+  urldate = {2023-04-24}
+}
+
+@article{deruijter2023,
+  title = {Ride-Pooling Adoption, Efficiency and Level of Service under Alternative Demand, Behavioural and Pricing Settings},
+  author = {{de Ruijter}, Arjan and Cats, Oded and {Alonso-Mora}, Javier and Hoogendoorn, Serge},
+  year = {2023},
+  month = may,
+  journal = {Transp. Plan. Technol.},
+  volume = {46},
+  number = {4},
+  pages = {407--436},
+  publisher = {{Routledge}},
+  issn = {0308-1060},
+  doi = {10.1080/03081060.2023.2194874},
+  urldate = {2023-12-02},
[…]
