+++ joss/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-RidePy provides fast computer simulations of on-demand mobility modes such as ridehailing or ridepooling, enabling both theoretical research into on-demand transit systems and the estimation of their operational properties in realistic scenarios. It strongly focuses on modeling the mobility service itself, rather than its customers or the environment, enabling the user to use any mobility demand model in conjunction with RidePy. This makes it possible to use RidePy in a broad range of scenarios. Through a combination of Python, Cython and C++, it offers ease of use at high performance. Its modular design makes customization easy, while the included modules allow for a quick start.
+RidePy provides fast computer simulations of on-demand mobility modes such as ridehailing or ridepooling, enabling both theoretical research into on-demand transit systems and the estimation of their operational properties in realistic scenarios. It strongly focuses on modeling the mobility service itself, rather than its customers or the environment, enabling the user to use any mobility demand model in conjunction with RidePy. This makes it possible to use RidePy in a broad range of scenarios. Through a combination of Python, Cython, and C++, it offers ease of use at high performance. Its modular design makes customization easy, while the included modules allow for a quick start.
 
 # Statement of need
 
@@ -31,9 +31,9 @@ The accelerating climate change and congested cities both call for an urgent cha
 In this context, a simulation framework should appropriately allow for vastly different system sizes and degrees of realism. The system size incorporates the number of simulated vehicles as well as the extent of the space they operate on: A small system may consist of a single vehicle serving a network of just two nodes, while an example of a large system could be a fleet of several thousand vehicles operating on the street network of a large city. The degree of realism may be varied, for example, by sampling requests from either a uniform distribution or recorded mobility demand, or by operating on a continuous Euclidean plane versus a realistic city street network. 
 Another option is to adjust the constraints imposed, such as the time windows assigned to stops, or the vehicles' seat capacities. 
 
-Finally, an on-demand mobility simulation framework should be fast, easy to use and adaptable to various applications.
+Finally, an on-demand mobility simulation framework should be fast, easy to use, and adaptable to various applications.
 
-A number of open-source simulation software projects are already being used to investigate on-demand mobility services. Some of them focus on microscopic modeling in realistic settings, through which concrete predictions for service operation are enabled, guiding urban planning. Prominent examples are MATSim [@horni2016], which performs agent-based simulations of individual inhabitants, and Eclipse SUMO [@lopez2018], a microscopic traffic simulator. Both rely on additional packages to model on-demand mobility, such as AMODEUS [@ruch2018] for MATSim and Jade [@behrisch2014] for SUMO.
+Several of open-source simulation software projects are already being used to investigate on-demand mobility services. Some of them focus on microscopic modeling in realistic settings, through which concrete predictions for service operation are enabled, guiding urban planning. Prominent examples are MATSim [@horni2016], which performs agent-based simulations of individual inhabitants, and Eclipse SUMO [@lopez2018], a microscopic traffic simulator. Both rely on additional packages to model on-demand mobility, such as AMODEUS [@ruch2018] for MATSim and Jade [@behrisch2014] for SUMO.
 
 FleetPy [@engelhardt2022], a recently released on-demand mobility simulation, is primarily aimed at realistic modeling of the interactions between operators and users, specifically incorporating multiple operators. While its technical approach is similar to ours, integrating Python with fast Cython and C++ extensions, the project is predominantly focused on applied simulations, although its framework architecture promises to allow for adjustment of the model detail level.
 
@@ -45,9 +45,9 @@ RidePy extends this landscape by providing a universal and fast ridepooling simu
 
[…]
