+++ joss/paper.bib
@@ -252,7 +252,7 @@
 
 @misc{ridepy-doc,
   title = {{{RidePy}} Documentation},
-  author = {{Felix Jung} and {Debsankha Manik}},
+  author = {Jung, Felix and Manik, Debsankha},
   year = {2023},
   url = {https://ridepy.org/},
   urldate = {2023-04-25},
@@ -261,7 +261,7 @@
 
 @misc{ridepy-github,
   title = {{{PhysicsOfMobility}}/Ridepy - Github},
-  author = {{Felix Jung} and {Debsankha Manik}},
+  author = {Jung, Felix and Manik, Debsankha},
   year = {2020},
   url = {https://github.com/PhysicsOfMobility/ridepy},
   urldate = {2023-04-25},
@@ -272,7 +272,7 @@
 
 @misc{ridepy-pypi,
   title = {Ridepy - {{PyPI}}},
-  author = {{Felix Jung} and {Debsankha Manik}},
+  author = {Jung, Felix and Manik, Debsankha},
   year = {2023},
   url = {https://pypi.org/project/ridepy/},
   urldate = {2023-04-25},
