### 2024-05-02 #6241 @fxjung
> @fxjung Please check this one here
>
>     * 10.1007/978-1-4842-2677-3_5 may be a valid DOI for title: Pytest

@diehlpk Unfortunately, this DOI is incorrectly suggested by the bot (see also my earlier comment on this in the pre-review [here](https://github.com/openjournals/joss-reviews/issues/6122#issuecomment-1841437649)). I was unable to find a suitable DOI for `pytest`.

Finally, I have just fixed some minor style issues in the manuscript and completed the **checklist**:

  - Version: `ridepy 2.7`
  - DOI: `10.5281/zenodo.11084287`
