+++ joss/paper.md
@@ -33,7 +33,7 @@ Another option is to adjust the constraints imposed, such as the time windows as
 
 Finally, an on-demand mobility simulation framework should be fast, easy to use, and adaptable to various applications.
 
-Several of open-source simulation software projects are already being used to investigate on-demand mobility services. Some of them focus on microscopic modeling in realistic settings, through which concrete predictions for service operation are enabled, guiding urban planning. Prominent examples are MATSim [@horni2016], which performs agent-based simulations of individual inhabitants, and Eclipse SUMO [@lopez2018], a microscopic traffic simulator. Both rely on additional packages to model on-demand mobility, such as AMODEUS [@ruch2018] for MATSim and Jade [@behrisch2014] for SUMO.
+Several open-source simulation software projects are already being used to investigate on-demand mobility services. Some of them focus on microscopic modeling in realistic settings, through which concrete predictions for service operation are enabled, guiding urban planning. Prominent examples are MATSim [@horni2016], which performs agent-based simulations of individual inhabitants, and Eclipse SUMO [@lopez2018], a microscopic traffic simulator. Both rely on additional packages to model on-demand mobility, such as AMODEUS [@ruch2018] for MATSim and Jade [@behrisch2014] for SUMO.
 
 FleetPy [@engelhardt2022], a recently released on-demand mobility simulation, is primarily aimed at realistic modeling of the interactions between operators and users, specifically incorporating multiple operators. While its technical approach is similar to ours, integrating Python with fast Cython and C++ extensions, the project is predominantly focused on applied simulations, although its framework architecture promises to allow for adjustment of the model detail level.
 
