### 2024-05-02 #6241 @fxjung
@diehlpk I've just fixed the typo in my repository (commit: https://github.com/PhysicsOfMobility/ridepy/commit/04aaa48bba2cbed97d519c804d95166b87860aa8)
