### 2024-05-02 #6241 @fxjung
@diehlpk Also, regarding the correct version of the software: I wasn't sure what the bot exactly needs---the version name is "ridepy 2.7", the git tag referring to that version is "v2.7" (see https://github.com/PhysicsOfMobility/ridepy/releases/tag/v2.7).
