### 2024-05-02 #6241 @diehlpk
> @diehlpk I've just fixed the typo in my repository (commit: [PhysicsOfMobility/ridepy@04aaa48](https://github.com/PhysicsOfMobility/ridepy/commit/04aaa48bba2cbed97d519c804d95166b87860aa8))

@fxjung the latest paper.md in the repo is used to generate the final paper. So you should be good.
