### 2024-05-03 #6241 @diehlpk
@fxjung the `paper.md` in the branch `joss` will be the final paper when the paper is accepted. So whatever `paper.md` is there once the editor in chief accepts the paper will be published.
