### 2024-05-03 9bc1615
Added MinimalPassengerTravelTimeDispatcher for Cython
