+++ src/ridepy/util/dispatchers_cython/cdispatchers.h
@@ -369,6 +369,208 @@ public:
   virtual ~AbstractDispatcher(){};
 };
 
+template <typename Loc>
+InsertionResult<Loc>minimal_passenger_travel_time_dispatcher(
+  std::shared_ptr<TransportationRequest<Loc>> request,
+  vector<Stop<Loc>> &stoplist,
+  TransportSpace<Loc> &space, int seat_capacity,
+  bool debug = false) {
+
+  double min_cost = INFINITY;
+  bool boolInsertEnd = true;
+  int best_pickup_idx = stoplist.size() - 1;
+  int best_dropoff_idx = stoplist.size() - 1;
+
+  // printf("------------ New request ------------\n");
+  // printf("Request origin: %i\n", request->origin);
+  // printf("Request destination: %i\n", request->destination);
+  // printf("Stoplist: ");
+  // for (auto i: stoplist) {
+  //   std::cout << i.location << ' ';
+  // }
+  // printf("\n");
+
+  for (int counter = 0; counter < stoplist.size() - 1; counter++) {
+    // printf("Counter: %i\n", counter);
+    Stop stop_before_pickup = stoplist[counter];
+    Stop stop_after_pickup = stoplist[counter + 1];
+    auto listResultInBetweenTest = is_between(space, request->origin, stop_before_pickup, stop_after_pickup);
+    if (std::get<0>(listResultInBetweenTest) == true && std::get<2>(listResultInBetweenTest) != 0) {
+        // printf("stop_before_pickup.occupancy_after_servicing %i\n", stop_before_pickup.occupancy_after_servicing);
+        // printf("seat_capacity %i\n", seat_capacity);
+        if (stop_before_pickup.occupancy_after_servicing == seat_capacity) {
+            continue;
+        }
+        double time_to_pickup = space.t(stop_before_pickup.location, request->origin);
+        double CPAT_pu = cpat_of_inserted_stop(stop_before_pickup, time_to_pickup);
+        double EAST_pu = request->pickup_timewindow_min;
+        if (CPAT_pu > request->pickup_timewindow_max) {
+            continue;
+        }
+        best_pickup_idx = counter;
+        // printf("Best pickup idx: %i\n", best_pickup_idx);
+        bool boolDropOffEnroute = false;
+        bool boolContinuePickUpLoop = false;
+        bool boolBreakPickUpLoop = false;
+
+        auto listResultInBetweenTestFollowingPickUp = is_between_2(space, request->destination, request->origin, stop_after_pickup);
+        if (std::get<0>(listResultInBetweenTestFollowingPickUp) == true && std::get<2>(listResultInBetweenTestFollowingPickUp) != 0) {
+            best_dropoff_idx = counter;
+            min_cost = CPAT_pu;
+            boolDropOffEnroute = true;
+            boolInsertEnd = false;
+            break;
+        }
+
+        for (int counter_drop_off = best_pickup_idx; counter_drop_off < stoplist.size() - 1; counter_drop_off++) {
+          // printf("Counter drop off: %i\n", counter_drop_off);
+          Stop stop_before_dropoff = stoplist[counter_drop_off];
+          Stop stop_after_dropoff = stoplist[counter_drop_off + 1];
+          if (counter_drop_off == counter) {
+              continue;
+          }
+          auto listResultInBetweenTestDropOff = is_between(space, request->destination, stop_before_dropoff, stop_after_dropoff);
+          //Changed from 2 to 1
+          
+          if (std::get<0>(listResultInBetweenTestDropOff) == true && std::get<2>(listResultInBetweenTestDropOff) != 0) {
+              best_dropoff_idx = counter_drop_off;
[…]
