### 2024-05-06 #6241 @danielskatz
👋 @fxjung - I've suggested a bunch of small changes in https://github.com/PhysicsOfMobility/ridepy/pull/260 - please merge this, or let me know what you disagree with, then we can continue the process and accept/publish this work
