### 2024-05-06 #6241 @fxjung
@danielskatz Thanks a lot for taking the time to read through our paper and improve it!
(Sorry for the curly brackets, they were generated by Zotero/better-bibtex.)

The only thing I'm still a bit unsure about (and have already been when writing it) is the sentence

> Accelerating climate change and congested cities both call for an urgent change in the way we move [@winkler2023].

To me, the beginning sounds a bit like we're trying to encourage accelerating climate change, although this is obviously a nonsensical interpretation. I don't know if there's a better way to phrase it.

But---I am not a native speaker. So if you think that this a non-issue, I'm completely fine with publishing it like this!
