### 2024-05-06 #6241 @danielskatz
I see your point.  An option would be to say
"The facts that climate change is accelerating and that cities are congested call for ..."
Would this work for you?
