### 2024-05-06 1bf1e8e
minor changes in paper and bib (#260)
* minor changes in paper

* minor changes paper.bib

someone seems to really like curly brackets :)

* Update paper.md
