+++ joss/paper.bib
@@ -1,13 +1,13 @@
 @book{behrisch2014,
-  title = {Simulation of {{Urban Mobility}}: {{First International Conference}}, {{SUMO}} 2013, {{Berlin}}, {{Germany}}, {{May}} 15-17, 2013. {{Revised Selected Papers}}},
-  shorttitle = {Simulation of {{Urban Mobility}}},
+  title = {Simulation of Urban Mobility: First International Conference, {SUMO} 2013, {Berlin}, {Germany}, {May} 15-17, 2013. {Revised Selected Papers}},
+  shorttitle = {Simulation of Urban Mobility},
   editor = {Behrisch, Michael and Krajzewicz, Daniel and Weber, Melanie},
   year = {2014},
-  series = {Information {{Systems}} and {{Applications}}, Incl. {{Internet}}/{{Web}}, and {{HCI}}},
+  series = {Information Systems and Applications, Incl. Internet/Web, and {HCI}},
   edition = {1st ed. 2014},
   number = {8594},
-  publisher = {{Springer Berlin Heidelberg : Imprint: Springer}},
-  address = {{Berlin, Heidelberg}},
+  publisher = {Springer Berlin Heidelberg: Imprint: Springer},
+  address = {Berlin, Heidelberg},
   doi = {10.1007/978-3-662-45079-6},
   abstract = {This book constitutes the thoroughly refereed proceedings of the First International Conference on Simulation of Urban Mobility, SUMO 2013, held in Berlin, Germany, in May 2013. The 12 revised full papers presented tin this book were carefully selected and reviewed from 22 submissions. The papers are organized in two topical sections: models and technical innovations and applications and surveys},
   isbn = {978-3-662-45079-6},
@@ -68,9 +68,9 @@
 
 @inproceedings{engelhardt2019,
   ids = {engelhardt2019a},
-  title = {Quantifying the {{Benefits}} of {{Autonomous On-Demand Ride-Pooling}}: {{A Simulation Study}} for {{Munich}}, {{Germany}}},
+  title = {Quantifying the Benefits of Autonomous On-Demand Ride-Pooling: A Simulation Study for {Munich}, {Germany}},
   shorttitle = {Quantifying the {{Benefits}} of {{Autonomous On-Demand Ride-Pooling}}},
-  booktitle = {2019 {{IEEE Intell}}. {{Transp}}. {{Syst}}. {{Conf}}. {{ITSC}}},
+  booktitle = {2019 IEEE Intell. Transp. Syst. Conf. {ITSC}},
   author = {Engelhardt, Roman and Dandl, Florian and Bilali, Aledia and Bogenberger, Klaus},
   year = {2019},
   month = oct,
@@ -82,12 +82,11 @@
 }
 
 @misc{engelhardt2022,
-  title = {{{FleetPy}}: {{A Modular Open-Source Simulation Tool}} for {{Mobility On-Demand Services}}},
+  title = {{FleetPy}: A Modular Open-Source Simulation Tool for Mobility On-Demand Services},
   shorttitle = {{{FleetPy}}},
   author = {Engelhardt, Roman and Dandl, Florian and Syed, Arslan-Ali and Zhang, Yunfei and Fehn, Fabian and Wolf, Fynn and Bogenberger, Klaus},
   year = {2022},
   month = jul,
-  number = {arXiv:2207.14246},
   eprint = {2207.14246},
   primaryclass = {cs, eess},
   publisher = {{arXiv}},
@@ -100,7 +99,7 @@
 }
 
[…]
