### 2024-05-06 #6241 @danielskatz
Congratulations to @fxjung (Felix Jung) and co-author on your publication!!

JOSS depends on volunteers and we couldn't be successful without your work
