### 2024-05-06 #6241 @fxjung
@danielskatz @diehlpk @xoolive @abhishektiwari @ixjlyons
Huge thanks to everyone here for all the work you put into our submission! Both the manuscript and RidePy itself improved a lot along this process, thanks to your feedback and suggestions!
